[roll] Roll fuchsia [netcfg] Initial interface name policy structure
Add in a new field to be read for configuration. The field is currently
optional and is represented as a default vector of size 0 when not
provided. For future CLs implementing the use of this configuration
field, a length of 0 will sustain current netcfg naming behavior and any
other lengths will use the new configuration based naming policies.

Create a struct for the MatchingRule and NameCompositionRule that will
will be extended in future CLs.

Ensure that the initial deserialization works for an interface naming
rule that has empty fields. Test will be a base for other features to
extend.

Original-Fixed: 135096
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/929969
Original-Revision: 844e8b2a16377aa232d620a0860182482b2d2a54
GitOrigin-RevId: ed357c4b0be90b90ee34f9fc38556baa2c8e6524
Change-Id: I3ca8ee246a5a4e616e974ad5b9f224cf869a05d8