[roll] Roll fuchsia [wlan] Migrate SoftMAC `SetChannel` to `WlanSoftmacBridge`.
This change migrates `SetChannel` from the `WlanSoftmac` protocol to the
`WlanSoftmacBridge` protocol and migrates from Banjo to FIDL in MLME and
the generic `wlansoftmac` driver.

As part of the migration, the `get_channel` binding and associated
device state has been removed, as no clients (i.e., MLME) use this
information and the channel was only cached in `set_channel` and did not
query the vendor driver.

Note that this change does **not** migrate `WlanRxInfo` and so Banjo and
FIDL conversions for channel types remain.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/929957
Original-Revision: 2b2a2069c65be6b14e5ac248008fcb3b91d95a90
GitOrigin-RevId: c8277e2135b5f954948b746ef9fbaa662d6db1a0
Change-Id: I156ece6582d6e3b63218b5a5f0c44759ac88118d