[roll] Roll fuchsia [uart] Allow more UART drivers on more machines
We've now seen the motmot and dw8250 drivers in RISC-V contexts, so we
allow their use there. In the case of dw8250, we further make it
available for all machines (we might as well for all of the ns8250
variants given their prevalence).

Original-Bug: b/299499151
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/936115
Original-Revision: fbb1e24f2fdcfb805b143e7908e5681947d3d513
GitOrigin-RevId: ae1750f048256f49de6e60c6fab2c81a221729ae
Change-Id: I2b5370d37b3d2a57a96f4b0176a5902cd77f0208