[roll] Roll fuchsia [pkg-cache] Protect cache packages from GC unconditionally
Previously, cache packages were activated in the dynamic index on
startup and were protected from GC as long as they remained active in
the dynamic index, i.e. as long as another package with the same name
(as indicated by the meta/package file) was not resolved.

Now, cache packages are still activated in the dynamic index on
startup, but they will be protected from GC regardless of if they
remain active.

This change was foretold by, and is in accordance with, the Open
Package Tracking RFC.

Original-Fixed: b/293332528
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/935191
Original-Revision: 06b75f0f16deb72f26ca390f9b8e487166b9c8a0
GitOrigin-RevId: 5f9eb88165f084d3f2ae8059d9862d2784498c95
Change-Id: I1a550ac090da62c2274ceb9c1c466005355d6139