[roll] Roll fuchsia [boot-shim][devicetree] Do not automatically overwrite UART driver
We update the devicetree boot shim's definition of InitMemory() to not
overwrite any pre-registered UART driver in the case where the chosen
matcher does not make a selection. This is useful in debugging and
bring-up scenarios where the UART driver might be hard-coded before
this function is called (and before we can even rely on the UART
matching logic to work).

Original-Bug: b/299499151
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/936116
Original-Revision: e57e0fe60f5b5636c08bf22adfca15a05221fe46
GitOrigin-RevId: 2f2b2bb1841d83b6be41f527042b1aefd4d7a127
Change-Id: If292fea9826a8c9d21c15bd7db91661d70d3bdcc