[roll] Roll fuchsia [kernel][arch] PagingTraits::Set() should imply accessed and dirty
Where the qualifiers make sense (which is implementation-specific), a
call to Set() should result in a present entry being updated as
accessed. There is no case in which we would want to create a new entry
but not mark it as such. If hardware is managing these bits, then there
is no penalty to doing so; if software-managed, then we avoid an
unnecessary trap that we would only handle later on to set these bits.
Similarly, if an entry is Set() as writable, then it should reflect as
'dirty' when that makes sense.

(Since we do not actually handle such traps in the phys environment,
this change also addresses the bug in which we end up in this situation
(seen recently on hardware) and needlessly fail to boot.)

Along these lines, we also remove `bool accessed` field from
PagingSettings and MapSettings.

Original-Bug: 91187
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/936213
Original-Revision: 80d56813bc7c6c4e5e4d5a1ce033f343950a22a2
GitOrigin-RevId: c10543fda911afc624fae4960a60e79860a681da
Change-Id: I5bef48b95c9c19904e31c2ea7c117a958816e2fe