[roll] Roll fuchsia [netcfg] Add DeviceClass MatchingRule
Use the DeviceInfo to determine which DeviceClass to assign the device.
The DeviceClass from `lib.rs` will be used for now and then for the
naming convention we can use the InterfaceType.

Add tests for Ethernet/Wlan/Ap cases where the correct device class
is specified in the list provided to the MatchingRule or
not provided.

A separate CL will create the infrastructure for checking for an
interface match across multiple `MatchingRule`s.

Original-Bug: 135094
Original-Bug: 135087
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/932207
Original-Revision: d6bb47f6b28c1b33d8b3dba61abb2efc13c7645c
GitOrigin-RevId: 6db7816f012353d4d311433300e5c7d8a1f07ad8
Change-Id: I0adc0f5e34526987ec403e5f9e0a147845d097e2