[roll] Roll fuchsia [superproject] Roll llvm-project/libc [libc] Enable remaining string functions on the GPU (#68346)
Summary:
We previously had to disable these string functions because they were
not compatible with the definitions coming from the GNU / host
environment. The GPU, when exporting its declarations, has a very
difficult requirement that it be compatible with the host environment as
both sides of the compilation need to agree on definitions and what's
present.

This patch more or less gives up an just copies the definitions as
expected by `glibc` if they are provided that way, otherwise we fall
back to the accepted way. This is the alternative solution to an
existing PR which instead disable's GCC's handling.

GitOrigin-RevId: 37da2e315590c7be0a8d74ad262e2d3d13651142
Original-Revision: 442f90f1a022bc18442e6451a058916da7eb0fce
Roller-URL: https://ci.chromium.org/b/8766379419926474289
CQ-Do-Not-Cancel-Tryjobs: true
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/936633
Original-Revision: 7de4651ce70d76f194dc7545f9ed6ea7bab7ca6a
Change-Id: I101162a66726671f9fe1ca561bce5a20a42502eb