[roll] Roll fuchsia [starnix] Return EMSGSIZE on sendmsg
When calling sendmsg on a zxio-backed socket, a return status of
ZX_ERR_OUT_OF_RANGE should be mapped to the EMSGSIZE errno, rather than
EINVAL, because it indicates that the provided input buffer was too
large to fit in the zircon object backing the socket (a channel in the
case of synchronous UDP).

This incorrect behavior was discovered when running the starnix LTP
tests against Netstack3, which happens to only support synchronous UDP
currently.

Test: starnix_ltp_tests
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/936332
Original-Revision: 1daa811d4110fa989d66bf08260d181a0f368b1b
GitOrigin-RevId: bfac85f6b25723f0cd4753dc3da9f7bf2e45dc82
Change-Id: I7a5eafecf6726a2b26ef1f0c890c1001abdfc7aa