[roll] Roll fuchsia [fidlc] Include resolved type of aliases in IR
This CL changes the FIDL JSON IR for "alias_declarations" to include the
fully resolved "type" in addition to "partial_type_ctor". I will look
into removing the latter after confirming nothing relies on it.

The motivation for this is we want to make a quick change to include
alias *definitions* in bindings (but not *uses*, which are much more
complicated). And the fidlgens only have logic to create a target
language type from the "type" schema, not "partial_type_ctor".

The partial type ctor concept exists because FIDL aliases were
originally more flexible, e.g. you could create an alias of "vector"
which is technically a type constructor, not a type. This is no longer
allowed, but even so, there are reasons you might want some more "raw"
representation of the right-hand side (see https://fxbug.dev/105758).
So we might have to keep "partial_type_ctor" -- or more likely design
something else to replace it.

Original-Bug: 7807, 105758
Test: fx check-goldens fidl
Test: fx test fidlc-test
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/936574
Original-Revision: e6037b5bb7c9800763fc5202c638d3d3cee35a85
GitOrigin-RevId: bc8205f75d53d9a7f091cd803971de0c4963fe91
Change-Id: I1262050fb29ce226523655cc10289421b0c4b7a7