[roll] Roll fuchsia [ffx][component][debug] Add `ffx component debug`
This command lets you attach to an existing component and debug it. This
command differs from the existing ways of attaching to components
because it uses the same user-friendly mechanisms from `ffx component`
to locate the component. For example, exactly the same queries work for
`ffx component debug` as work for `ffx component show`.

Original-Bug: 306672577
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/935179
Original-Revision: 98424c482fbad76242140ccb91b4c788f33bc796
GitOrigin-RevId: c4f5f42419290aeb7200e52747f9a2663a17fd4c
Change-Id: I47f22367dc40d7f169aa92277f17e06a68cbd969