[roll] Roll fuchsia [kernel][vm] Skip redundant mark accessed
Skips the rest of the mark accessed logic if the page is already in the
target queue. The added kcounters show this short circuit can happen up
to 70% of the time.

MarkAccessedDeferredCount is also moved out of the header as there is
no benefit to it being there since it is too large to be inlined at
call sites anyway.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/936552
Original-Revision: dad5f61821ab8c5952f5e5b04a76430a1871f16a
GitOrigin-RevId: 7bcc39e7096a24796c7ce545dacaa5d6fa58d0d3
Change-Id: I8ba97c5bc120db57bc172017cd677a1eaefc8301