[roll] Roll fuchsia [netcfg] Add TopologicalPath MatchingRule
Intake a pattern from the provided configuration and determine if
that matches against the device's topological path. If the string
is a match, this will constitute a match.

Add tests for patterns that match the provided path and do not.

I found out a way to make the pattern verified for correct formatting on
consumption of the configuration, which is very helpful for avoiding
run-time formatting problems. This functionality is tested in the
configuration consumption step.

Original-Bug: 135094
Original-Bug: 135087
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/933436
Original-Revision: 71d041f16329de27a6051b65499c8fcf23ce4b43
GitOrigin-RevId: e03b1a518b605b6230ff14551d99472a1333637b
Change-Id: I2f8eb17789456a7050f79635d3c0cdad7eb566e3