[roll] Roll fuchsia [trace][rust] Add TraceFuture for tracing futures
TraceFuture adds a duration event when a future is created, dropped, and
every time it's polled. The duration events are connected by flow
events. The "created" duration event unfortunately doesn't encompass the
creation of the wrapped future since the wrapped future must be
constructed before the TraceFuture.

The purpose of the TraceFuture is to aid in tracking the execution of
async rust code:
 - A future passed to Task::spawn may be first polled on a
   different thread from the one it was created on.
 - A future may be polled multiple times across multiple threads.
 - A future may be dropped before becoming Ready.

The main advantage of TraceFuture over async events is that TraceFuture
shows the synchronous steps of an asynchronous task. Async events are
rendered separately from synchronous duration events which makes it
difficult to associate duration events with async events if multiple
async events are overlapping.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/934058
Original-Revision: 5059b5bbb965ec8d8608c12ad5952ed04c549d11
GitOrigin-RevId: 5314d57e6bd0aa0aabadef6db4737b407f7ffd0f
Change-Id: Id147cdd8b77a01c213fc31203b8d752bb2648cf6