[roll] Roll fuchsia [devfs] Remove multiplexing from base path
Removes the multiplexing at the base devfs path to just connect to the
device protocol. Keeps the node as a default when things are running
with dfv1 since there are a lot of fvm tests that use devfs paths with
fds.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/934807
Original-Revision: 4419e98200cfc3ad65172d1bf141f186edd1397e
GitOrigin-RevId: 8a279df94460eab89c78e3bfb686811ae0235b5a
Change-Id: Ic29644c0d0e93fc3897295bbfb09856c9f18af85