[roll] Roll fuchsia [installer] Add missing test dependency
This test was missing a dependency on test data. We did not notice this
initially, as this test was run amongst other tests that depended on the
data, indirectly ensuring that it would be available to our test. When
run in a more isolated set of tests though, this test would consistently
fail.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/937686
Original-Revision: 5158fdfd05e4193e7a501ab22a3832c2f867f4b2
GitOrigin-RevId: 87e674951ade15f9e8a44aacd5ea382af9409345
Change-Id: Ie774ded48f59dbc89fb2c2db446642b2201bccff