[roll] Roll fuchsia [qemu] Use an SMBIOS v2.1 entry point when booting x86 guests
As of QEMU version 8.1.5, the default SMBIOS entry point type is the
v3/64-bit kind instead of the v2.1/32-bit kind. We only support the
latter at the moment.

This QEMU patch introduces the (undocumented??) pc-specific property of
`smbios-entry-point-type` that allows us to override this default:
https://patchwork.kernel.org/project/qemu-devel/patch/20211026151100.1691925-4-ehabkost@redhat.com/

We set this to unblock a future roll to (or past) QEMU version 8.1.5. We
might consider supporting SMBIOS v3 in the future, but no need to block
the roll on that now.

This property was tested locally with both v8.0.3 (the current version)
and v8.1.5: we successfully booted with an SMBIOS v2.1 entry point in
each case.

Original-Fixed: 134894
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/936612
Owners-Override: Adam Barth <abarth@google.com>
Original-Revision: 96ef772291481d30da84420cb3e92ffd74efd6a6
GitOrigin-RevId: d04172faaf34ea4f6f902ddb945671a592fbd343
Change-Id: I8e7ee0cde28d6beed0b2dd9ecacd1dcaf80e2e52