[roll] Roll llvm-project/libc [libc] Compile the GPU functions with '-fconvergent-functions' (#70229)
Summary:
This patch simply adds the `-fconvergent-functions` flag to the GPU
compilation. This is in relation to the behaviour of SIMT
architectures under divergence. With the flag, we assume every function
is convergent by default and rely on the compiler's divergence analysis
to transform it if possible.

Fixes: https://github.com/llvm/llvm-project/issues/63853
GitOrigin-RevId: bde11aaf8562d7296b87b278103014b9d513437a
Original-Revision: 37d1e79d9dfbbffdabf540eef62987bba477219c
Change-Id: Ibaec4b2c67bc4007109bc88b66299ad44ee9da8a