[roll] Roll fuchsia [superproject] Roll llvm-project/libc [libc] Compile the GPU functions with '-fconvergent-functions' (#70229)
Summary:
This patch simply adds the `-fconvergent-functions` flag to the GPU
compilation. This is in relation to the behaviour of SIMT
architectures under divergence. With the flag, we assume every function
is convergent by default and rely on the compiler's divergence analysis
to transform it if possible.

Fixes: https://github.com/llvm/llvm-project/issues/63853
GitOrigin-RevId: 5256e177bcf47a95b72f2afc750bf5d8c5df4ec5
Original-Revision: 37d1e79d9dfbbffdabf540eef62987bba477219c
Roller-URL: https://ci.chromium.org/b/8766198525117424161
CQ-Do-Not-Cancel-Tryjobs: true
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/937789
Original-Revision: e2485ce0cfce15a0023276a29770234e4def9452
Change-Id: I78cc557e0dbb81159e02978df158aa03065bd0f5