[roll] Roll fuchsia [kernel][vm] Avoid excessive scanning
Changes the way the access_scan_period is calculated and used to avoid
redundant scans. The intention of the regular access scanner is that
there is at most access_scan_period between scans, but the existing
implementation had a scenario where if did a routine scan at time A,
and then waited till time A+period, if a scan from the page queues
were to happen just moments before (i.e. A+period-delta) then we would
still rescan.

This change corrects this issue and ensures that the periodic scan
only occurs if nothing else triggered a scan for at least the period.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/937853
Original-Revision: 379e65098c5c8de279b911ae417d9d883fcccb66
GitOrigin-RevId: 93cb71c15ce221bf4471009746a3a2a8d280a60c
Change-Id: Ie6d74c723ccc4e6c3939b36ec87bc978248927bd