[roll] Roll fuchsia [display] Remove id from sysmem debug name
The id is a separate field, and by not including the id in the name we
get the desired porperty that a sysmem(1) name not vary per instance.
In sysmem2 we'll separate name and client_code_id to make it ok to vary
per instance in the name as long as a client_code_id is also set that
doesn't vary per instance.

BUG: b/284073556
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/938443
Original-Revision: 32b1b3c8512762c46eaa53201f129f91b3e0d7f6
GitOrigin-RevId: 300911ea8b71e846ee1a42f594eac6639318ecfd
Change-Id: Ifbeec23a1a91ddfb388e7b9acbd55fbbaba9937d