[roll] Roll fuchsia [assembly] Product configs can contain lists of files
Adds support for lists of product-provided files in the product config.

The GNI change is necessary so that we list each file as a GN input and
rebuild when one of those files change.

The BZL change is necessary so that we copy the input files into the
bazel sandbox before running assembly.

Original-Bug: b/299808720
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/937661
Original-Revision: 2e614e605639c2665c106b86972f817c1054fc51
GitOrigin-RevId: b367c2d1b5784178a0b7bf77b74f0bc820a1620e
Change-Id: Ief58c17b0bbacfc5e611579c3d2bf2de0936d3a3