[roll] Roll fuchsia [hierarchy] Do not return empty hierarchies when the properties weren't matched
I'm honestly not sure on the history of why we built it this way, but
looks like it was intentionally built. However, this is unexpected and
messes up with fetch-retry approaches which expect a request to
succeed only when the full selector was matched, not on partial matches.

Additionally this should result in smaller snapshots when the data
we want isn't there. Rather than including all the path to the data,
we simply don't include the data. This can also translate into using
less memory, but we'd need to measure to see the impact in practice.

Original-Fixed: b/303732863
Run-All-Tests: True
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/938233
Original-Revision: 8b0415dfeb1ac526fb409607df3d7f74a2b124ae
GitOrigin-RevId: e860705474bbfa1fd451b8e7a579b91e2535c2f1
Change-Id: I4c980e91e66fc9603474d257632629a73e7fd807