[roll] Roll infra/recipes [git] Fix ordering of commits in diff-tree command.
The base commit should come first in the command before the commit
we're trying to get changed files for. See go/bbid/8767320495216881969
for an example of how it was getting the wrong list of changed files
because it was listing a deleted file instead of the renamed one.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/infra/recipes/+/938617
Original-Revision: 6a1c38f620113cb2faf829c9eec1d6f72590669f
GitOrigin-RevId: 0fb101264a401f601c08058277f622aad1127dad
Change-Id: I48bdda90aa5a496b94233b94721320ccb572745f