[roll] Roll llvm-project/libc [libc] Optimize the RPC memory copy for the AMDGPU target (#70467)
Summary:
We previously made the change to make the GPU target use builtin
implementations of memory copy functions. However, this had the negative
effect of massively increasing register usages when using the printing
interface. For example, a `printf` call went from using 25 VGPRs to 54
simply because of using the builtin. However, we probably want to still
export the builitin, but for the RPC interface we heavily prefer small
resource usage over the performance gains of fully unrolling this loop.
For NVPTX however, the builtin implementation causes the resource usage
to go down (36 registers total for a regular `fputs` call) so we will
maintain that implementation.

I think specializing this is the right call as we will always prefer the
implementation with the smallest resource footprint for this interface,
as performance is already going to be heavily bottlenecked by the use of
fine-grained memory.

GitOrigin-RevId: dfc76fe8251ac3c44dd578c129d15004f99acd1c
Original-Revision: c36006f5ebe59c5571d1035e7701ff8564fd95b4
Change-Id: Ic44bbb5cb6380156a5b231e3002971d60c79d598