[roll] Roll fuchsia [superproject] Roll llvm-project/libc [libc] Optimize the RPC memory copy for the AMDGPU target (#70467)
Summary:
We previously made the change to make the GPU target use builtin
implementations of memory copy functions. However, this had the negative
effect of massively increasing register usages when using the printing
interface. For example, a `printf` call went from using 25 VGPRs to 54
simply because of using the builtin. However, we probably want to still
export the builitin, but for the RPC interface we heavily prefer small
resource usage over the performance gains of fully unrolling this loop.
For NVPTX however, the builtin implementation causes the resource usage
to go down (36 registers total for a regular `fputs` call) so we will
maintain that implementation.

I think specializing this is the right call as we will always prefer the
implementation with the smallest resource footprint for this interface,
as performance is already going to be heavily bottlenecked by the use of
fine-grained memory.

GitOrigin-RevId: 03dc20a4844b70f8df342c80159df118e3466484
Original-Revision: c36006f5ebe59c5571d1035e7701ff8564fd95b4
Roller-URL: https://ci.chromium.org/b/8766046456596995601
CQ-Do-Not-Cancel-Tryjobs: true
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/938500
Original-Revision: f6e45e1cc34ff6372f3412cee44efbcfebb1083a
Change-Id: I45a24ccdadf9285111de69860ed5cd997a314b40