[roll] Roll fuchsia [superproject] Roll llvm-project/libc [amdgpu][openmp] Avoiding writing to packet header twice (#70695)
I think it follows from the HSA spec that a write to the first byte is
deemed significant to the GPU in which case writing to the second short
and reading back from it later would be safe. However, the examples for
this all involve an atomic write to the first 32 bits and it seems a
credible risk that the occasional CI errors abound invalid packets have
as their root cause that the firmware notices the early write to
packet->setup and treats that as a sign that the packet is ready to go.

That was overly-paranoid, however in passing noticed the code in libc is
genuinely invalid. The memset writes a zero to the header byte, changing
it from type_invalid (1) to type_vendor (0), at which point the GPU is
free to read the 64 byte packet and interpret it as a vendor packet,
which is probably why libc CI periodically errors about invalid packets.

Also a drive by change to do the atomic store on a uint32_t
consistently. I'm not sure offhand what __atomic_store_n on a uint16_t*
and an int resolves to, seems better to be unambiguous there.

GitOrigin-RevId: 36cf636257860737e4c0b10c23ecce3679f368a2
Original-Revision: 1f1a65436e171086b4a4116a227b35049375bb0b
Roller-URL: https://ci.chromium.org/b/8765780518576579281
CQ-Do-Not-Cancel-Tryjobs: true
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/939204
Original-Revision: 0bd47944d2464bcf9ec3706fe23ccb87ac1cef75
Change-Id: I653f024cfdfa861535952a9f3185535d54ec79b3