[roll] Roll fuchsia [netstack3] Make "bound in both stacks" explicit
A dual-stack enabled socket listening to the IPv6 any address is
considered to be bound in both stacks. Encode this explicitly as a
third-variant in the `DualStackListenerIpAddr` enum, which forces the
`BothStacks` case to be handled when interacting with listener sockets.

While here, adjust the `DualStackListenerIpAddr` to hold
`ListenerIpAddr` rather than it's own custom `DualStackIpAddr`. This
brings it more inline with other dualstack data types.

Original-Bug: 135204
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/938827
Original-Revision: 63a4427fa48ecbf24d720b6836f70a56704ceae5
GitOrigin-RevId: ee25805f46ff5b2e1faaa8ea9783e22e78fb9b6f
Change-Id: Ia2fa82b21293b83c9f6997d2f05facbf55df366a