[roll] Roll fuchsia [ld] Support IE model TLS access
This adds the support for the IE dynamic relocs and tests them.
This includes tracking DF_STATIC_TLS and insisting that the IE
reloc types are only used in a module that set it.  In the
startup dynamic linker, it's actually no problem to just let
any module use these relocs since by definition they are all in
the IE set.  But it's the static linker's responsibility to
identify a IE-using module with DF_STATIC_TLS so that dlopen
can refuse to load it, so it makes sense to insist that the
static linker did the right thing even in the startup context.

Original-Bug: 128502
Cq-Include-Trybots: luci.turquoise.global.try:run-postsubmit-tryjobs
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/934913
Original-Revision: e72475bcd1660943c83ce9d0a7c80f68b1aab683
GitOrigin-RevId: 3c712655b41b13e19f4a274008c337221c0317bd
Change-Id: Ibe0da6451b4ef22b136db7d30cbe0021404bcd46