[roll] Roll fuchsia [ld] Fix ld.so ABI verification to use right variant
For ld-startup, the ABI verification is applied to each variant
that is used in one of the tests.  For ld-stub, the ABI verification
needs extra plumbing to use explicit variant selection so that it
will match the context where ld-stub gets built.

Original-Bug: 130483
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/939014
Original-Revision: bdb3004764c2c734b55a456828b0bcaf67bb4ee8
GitOrigin-RevId: c60b8decbfb74670d19887477699cf0cfe1c9a3f
Change-Id: I15d002706071593b0bc147b3e07fb7fffbbd02ce