[roll] Roll fuchsia [time] Set SIGNAL_UTC_CLOCK_LOGGING_QUALITY
This change starts signaling that the UTC clock is of logging
quality. This is signaled when we either read something from
the RTC, or something from an external source.

The resulting UTC clock is adequate for debug logging, but nothing else.

Original-Bug: b/308214128
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/938941
Original-Revision: 203261718414b1b193f234d2c26b97c3921accb9
GitOrigin-RevId: 1e7db662d0597845cfbe5879618d708185513403
Change-Id: I6bc791e21cbc885ae2e63510d30e4b4d35008ac0