[roll] Roll fuchsia [netstack] Populate socket inspect data for test
Formerly, the ArchiveReader would return an empty DiagnosticsHierarchy
if a snapshot was taken for a selector which did not correspond to
Inspect data. After a recent change, it now returns an
Option<DiagnosticsHierarchy>. That change revealed that the
netstack-sys-integration-test did not retrieve Socket Inspect data in
the relevant tests. Create a UDP socket to ensure that such data exists,
and then remove a special case that had been added to account for the
missing Socket data. While I'm here, remove an unnecessary reference
operator and add a comment documenting a subtlety of one of the test
helpers.

Tested: fx test netstack-sys-integration-test
Original-Fixed: 135964
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/939896
Original-Revision: e1b9d756782af476ca0be0cba45da10423660ade
GitOrigin-RevId: 3561be058c8f9f47d3ddd812e7bc7f30724994e5
Change-Id: Ia3b1bba602c155c19ce7e31e15367401847546cd