[roll] Roll fuchsia [bazel_sdk] allow fuchsia_cc_driver to specify binary name
Adds an attribute to fuchsia_cc_driver which allows users to specify the
name of the built driver. For now we keep the default as lib<name>.so
which is different from in-tree until we can migrate our OOT repos to
match the default that we have in-tree.

BUG: b/305314198
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/939989
Original-Revision: 54855530fbc9d62e77588c30259aae7a637ecbf2
GitOrigin-RevId: d64167f3b19379a1c7c75268e3f69cbda9803647
Change-Id: I2e1e91670089c86c4a1ff0b0c8d1a3e673164a7b