[roll] Roll fuchsia [ui] Route GraphicalPresenter to session
fuchsia.session.scene.Manager is deprectaed and we intend to remove it
eventually. Making this change now allows us to avoid adding it to the
partner SDK, for products which currently need it and are moving OOT.

Remove references to an obsolete bug which said we should remove the
GraphicalPresenter routes. It is no longer relevant.

Part of "Refactoring Session and Window Manager Components" migration

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/940112
Original-Revision: ed41669f1a43d638d3cdd302ef0744ef5c287ffc
GitOrigin-RevId: 1142dc4d074eab73caa7e4eb761f61b104612005
Change-Id: I883a782ca3fc3bd60e788846fcb979d877f81b18