[roll] Roll fuchsia [testsharder] Fix computation of duration of multiplied shards.
We recalculate the shard timeout when creating multiplied shards
because if runs are not specified, then we need to fill up the shard up
to the target duration with the multiplied tests and then calculate the
timeout based on the number of runs we determine. However, in the case
where runs are already specified, the duration of running all multiplied
tests may exceed the target duration. In this case, we should just use
the already calculated timeout from sharding the multiplied tests the
first time since no new test runs will be added.

MULTIPLY: `[
  {
    "max_matches": 500,
    "name": "starnix_gvisor_tests",
    "total_runs": 3
  }
]`
Run-All-Tests: true
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/939879
Original-Revision: 8b8b067ee57a9434c49a28af14323310af1b4c34
GitOrigin-RevId: 2f00296ad450b2370bda37929d003035c63f06b6
Change-Id: I14aecce4f72836521f191d16413c5ae8e54e9411