[roll] Roll fuchsia [kernel] Introduce `kernel_extra_deps` GN arg
This arg is a list of GN labels comprising additional dependencies of
the kernel proper. This can be useful - in a prototyping or 'vendor'
capacity - for injecting new instances of subsystems that the kernel has
defined modularly (e.g., pdev drivers or k commands).

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/939987
Original-Revision: 0c80d66dfe568326c54280bc328f037dd8360ab9
GitOrigin-RevId: ce8431a5115fb04499a8f2d330213514354662ce
Change-Id: If604ecc7a25e4d93202de66d2f6d0729de131818