[roll] Roll fuchsia [debugger] Add OS info to connection.
Exposes the OS type in the IPC reply message from the debug agent, and
exposes this in various places.

Adds the platform and CPU architecture to the "status" output.

Updates the processing of the HelloReply message to allow versioning as
long as the version is set for the "version 0" deserialization. This
required changes on both the debug_agent and zxdb side to properly allow
versioning on both ends.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/939878
Original-Revision: a977ef4636c7a262bd882a7879f21e4a8226760a
GitOrigin-RevId: f6167a6a8f4a0d52ed41aab0d7954be695ea3946
Change-Id: I03c2a04331b46bb6cb4ecb811fb4ec31d851b202