[roll] Roll fuchsia [time][doc] Update the UTC clock documentation
The linked bug has updated the UTC clock behavior to include
an extra clock state and an extra clock signal on state entry.

We introduced this behavior for some limited testing use cases,
but we still want to communicate that the clock has changed,
and that there are new signals to watch.

Original-Bug: b/295537827
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/938698
Original-Revision: 87adee6030a5a9bc154a8ddaf91ca7209d597162
GitOrigin-RevId: 4c5809ce78f7034e89da41ab6603de8dff001ccf
Change-Id: I8b944912a227ed5e02453f7b4314ae65a754496c