[roll] Roll fuchsia [starnix] Usercopy routines using exceptions
This implements a usercopy library for copying memory in/out of
restricted address ranges that translates faults into returns with
errors. This is intended to be used in Starnix in direct memory mapping
mode to catch faults when accessing restricted memory in normal mode
execution such as syscall implementations.

This library uses an exception handler on the current process to catch
page faults within the usercopy routine with a faulting address in the
configured restricted range. Only x86_64 is supported for now.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/917566
Original-Revision: 950dfc909bff87ff1be80ce75ed723cf65d41950
GitOrigin-RevId: fd3c99db45757bfd0c95fe6ae512780987d20647
Change-Id: I92183d89f67d66b67b26dfcfc4d87ec93766d0bc