[roll] Roll fuchsia [starnix] Remove unused `zx::Job`
This job was previously used to hold all (child) Linux processes that
Starnix managed. With the move to use a restricted executor (backed by
`zx_process_create_shared` and family), the job was unused.

Note that `zx_process_create_shared` creates processes in the same job
that the shared process (e.g. Starnix) is a direct descendant of.

commit 429db297f0f5bce1e9dce86b019ea37eb909ae5e removed the use of the
job but forgot to remove the job itself.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/940198
Original-Revision: 3dfedf5891122ee7938921f46c5596ddcf037596
GitOrigin-RevId: c44615d82ce186a7b8975fb72d874950abd655c2
Change-Id: I0c0d81f143cc7c5aada5f8597518004b65d36147