[roll] Roll fuchsia [debugger] Handle overlapping ELF libraries.
When searching for loaded ELF libraries, GetElfModulesForProcess() would
avoid adding regions that overlapped with a previous ELF module (as
explained in the long comment in that function).

But this logic didn't account for modules found using the dl_debug_addr
method earlier in the code, and the "end" address of that module would
never be accounted for. This is a problem only when the mapping found
using the dl_debug_addr (the main exe) is the first mapping in memory
(normally not the case on Fuchsia) which is why we never noticed.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/933195
Original-Revision: 33065f22ee6a833807cffee061d90313d44f0ca9
GitOrigin-RevId: 716782e8579618ab0a2ca6e7bc35419281a26b43
Change-Id: Ieae95e2da5d46443d6ed570d2002f6f8b6ffe9ae