[roll] Roll fuchsia [testrunner][seriallistener] Use LineWriter for test output.
Test output and emulator output are both written to stdout and can
interleave with each other. We use a LineWriter to try to at least write
one complete line from each writer to stdout before writing from the
next writer. However, the test output writer was only using the
LockedWriter associated with stdout instead of the LineWriter. This fix
may or may not help with the attached bug, but it seems an issue with
what's getting written to stdout, so we can submit this and see if it
helps.

Original-Bug: 135386
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/940253
Original-Revision: 880ecfbbf56046bc5a4e0e7cb7859be4bb20766b
GitOrigin-RevId: 0405aff02b430d80efcdaa8699b28815e285d22d
Change-Id: I32ea5fc376df19d20fb3c8c6af29780df9f2dea5