[roll] Roll fuchsia [fidl][fuchsia-controller] Simplify FIDL lookup.
When looking up FIDL files programmatically, it is easier to look them up based
on the library name. This removes the need for all_fidl_json.txt from
fuchsia-controller to find FIDL files. So far a future drawback will be when
users eventually try to use a library that will have a duplicate name to another
library. This scenario is not supported.

Future work could be to handle setting up FIDL IR links on a per-target
basis, but this merits further discussion (beyond the scope of this
change).

Original-Bug: b/301178571

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/940058
Original-Revision: 1f99b5efb32031f7ae5ccd96fc3d7db5b8d836ce
GitOrigin-RevId: a787b8b771575b364bef2770fbcf6911d51502de
Change-Id: I8647189901a07a2462f746d4cbec03324d372781