[roll] Roll fuchsia [swd] System tests do not depend on bin/clock
Instead of using bin/clock to get the monotonic time, we now use ffx
target get-time. In order to ensure that ffx is compatible with the
older builds, we use ffx from the build itself. To ensure that we are
always using the correct ffx for the specific device, the Reconnect()
method now takes a 'build' so that after a reboot we reinitialize our
connection using the ffx from the 'build'.

Because older ffx version are unable to restart the daemon for newer ffx
versions, the Reconnect() is a bit more complicated because it has to
manually stop the daemon using the newer ffx before starting the daemon
with the older ffx. Because infra does not support mdns, we have to
explicitly add the ffx target after starting the daemon.

Original-Bug: b/304592348
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/938633
Original-Revision: 5951aecf31c8a86e58df3f99100ec55eb35bd084
GitOrigin-RevId: 61c92fa1aaa64d5eff96ec4685c98cce741a5963
Change-Id: I78b1a6e39a0fa04bfc69d3ee23c52b0f463e2d0a