[roll] Roll llvm-project/libc [libc] Allow hermetic timing if the `clock` function is built (#71092)
Summary:
This patch fixes some code duplication on the GPU. The GPU build wanted
to enable timing for hermetic tests so it built some special case
handling into the test suite. Now that `clock` is supported on the
target we can simply link against the external interface. Because we
include `clock.h` for the CLOCKS_PER_SEC macro we remap the C entrypoint
to the internal one if it ends up called. This should allow hermetic
tests to run with timing if it is supported.
GitOrigin-RevId: 5c424746ac2bcc0b973bd812b65c8b07d1df03b3
Original-Revision: 1252ff83d1864e41579fb8b2dab905b4d44f4add
Change-Id: I4e2f8ea14cde663d07823c2676eccb96f7ba219e