[roll] Roll fuchsia [swd-e2e-tests] Max size should be on system image size
The max size we are testing with is the system image or update images
size, which should exclude the update package size, since that cannot
get garbage collected during an OTA.

This also restructures how the size is calculated by moving the
computation into the Package types to deduplicate code.

As part of this, it changes the argument from `-max-ota-size` to
`-max-system-image-size`.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/940444
Original-Revision: 3fe35cd6ae39264a201becae2145b5436550fa15
GitOrigin-RevId: 9192f11317aa0193207e2e27c23ac8c50737942f
Change-Id: I47fbef1610c2afaf06622091fa1f3eb53c977894