[roll] Roll fuchsia [devshell] Use mac-arm64 binaries when on arm64 mac
We have brought up all but a handful of (mostly unused) tools that
worked on amd64 macs with arm64 prebuilts.  Let's allow them to be used
directly so people with such machines no longer have to run builds under
Rosetta.

This is also expected to fix the incremental build failures on the m1
bot in FYI, since now it should be able to find the correct
fuchsia-vendored-python python3 binary on PATH, since it'll look in the
same place that we've actually populated the prebuilts.

Original-Bug: b/283475308
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/941173
Original-Revision: 1319ff1277c9202ea79ae72d2738ef9c91ddc085
GitOrigin-RevId: 5f90a7d339243c0d4672e2613fb105176040f149
Change-Id: Iba88a7d1a7d370882472c83db83dc191fda29449