[roll] Roll fuchsia [netcfg] Name already exists from naming policy
DynamicNameCompositionRules will need to know information about existing
interfaces to know whether the name needs to be altered, such as in the
case of finding a unique number to normalize a MAC address.

Accessing the interface names during name generation will make it
easier to determine whether an interface with the same name already
exists.

Test added to ensure that an error is surfaced when an interface with
the same name already exists in the Config.

Original-Bug: 135087
Original-Bug: 136013
Original-Bug: b/309010767
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/939876
Original-Revision: ccc082fa26aab6d9ed8a9ddaa1fee8de34072e4c
GitOrigin-RevId: 82d1db23c829e71497ce4f276fdcc45f8d9a127e
Change-Id: I291323fc0c5c9fc8fac27d133127b7c92a4e8037