[roll] Roll fuchsia [config-netstack] Isolate unused endpoints in test
All the endpoints in the configurable-netstack-test are on the same
broadcast domain, which means in the connect_to_server test, all the
endpoints that are used in the other test cases are also receiving the
traffic from the first test's client and server. This causes a lot of
unnecessary processing and log noise, so instead place them in an
isolated network (that sees no traffic).

Original-Bug: 136090
Multiply: configurable-netstack-test
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/941475
Original-Revision: 8847c6003a38e7b3fcea7d2411d37505099fb63f
GitOrigin-RevId: 43e99b8948872d5b34da3fa3a1963d2a2bf9d632
Change-Id: If22d642dea95a46b1be128fb9c46d747f319a0e4