[roll] Roll fuchsia [assembly] Move kernel into its own platform input bundle
Move the kernel out of the legacy input bundle, and into its own
platform AIB, using a special template that can use the metadata-based
contract with the kernel portion of the build to locate the path to the
file.

This also changes assembled_system() to create a 'custom kernel AIB' if
a non-standard kernel image is passed to the assembled_system() GN
template.  A new option is added to product assembly to receive the path
to that AIB and use it in lieu of the standard kernel image in the
platform AIBs.

Original-Bug: b/308686895
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/940438
Original-Revision: 64358272ab0effecf2f38cf3209b2df8c4ee237e
GitOrigin-RevId: e0d2452c7d35dead7d712509e319d0245312854a
Change-Id: Ibab9b293c56a4a1eebdbce32c0a39922b036609e