[roll] Roll fuchsia [zxdb] Speed up module loading when debugging tests
When `ffx test` runs tests, we end up running the same process many,
many times, once for each test case. This process was painfully slow
because we were spending a lot of time reload the same modules over and
over again.

We have a mechanism for caching modules, but we were hitting a
pathological behavior and evicting the modules from the cache
immediately before they were needed because we discover the modules for
a process in two steps: first we get notified of two modules (presumably
the vDSO and the loader) and then we get notified of a dozen more
modules (presumably the actual modules for the test). The first
notification was causing us to dump the remaining dozen from the cache.

This CL increases the lifetime of modules in the cache by preserving all
the modules from the previous instance of the process. With this change,
we are able to keep all the modules alive in the cache long enough for
the subsequent process to fully load, making the whole test run
massively faster.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/941596
Original-Revision: b87ef20efc9450bd7a69a0cfaa99ac3b5029bd50
GitOrigin-RevId: eb6cb33feab12acc59ad6a2c6235f8a7c743587b
Change-Id: I0ccc269ac494c367811b187639f82a2a36563a08