[roll] Roll fuchsia [kernel][arch] Introduce global support to the paging traits
This change adds support for global pages in the libarch paging traits.
This will be used in the eventual mapping of the kernel.

Because it was low-hanging exercise and otherwise still planned for, we
also enable global mappings in the phys environment and update the
virtual address loading test to map its test ELF image as global. This
is accomplished by adopted the policy that AddressSpace::Map() will
always fix up the input settings to set global when mapping in the upper
address space.

Original-Bug: 91187
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/900444
Original-Revision: 900f0a91c57f57f2177f14c8440f7f46adfb2d38
GitOrigin-RevId: 6fbd25ca39610a6e3a786a751d57a34033b13b63
Change-Id: Ic1ad4f4dc601a2fe102943243c4c991852484f93