[roll] Roll fuchsia [swd-e2e-tests] Fix adding blobs to system image and update images
This makes sure we always include the rest of the blobs referenced by
the system image or update images packages. Previously we would only
include them for consideration on the first pass, but would not
afterwards if the generated package was too large. That would ultimately
lead to us exceeding the physical capacity of the device.

This patch fixes the issue by making sure we always pass down the
additional blobs to `addRandomFiles` so they're included when we compare
the total size of the blobs to `maxSize`.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/941674
Original-Revision: ebaa085863fc526d2708da52f986248947d5eaa7
GitOrigin-RevId: ee32adf7af3a1002e4c456eda469115c61afc3d8
Change-Id: I92f48c3f3274941a4caed4f262ac9a8216f18951