[roll] Roll fuchsia [fidl][cpp] Use --assume-filename to speed up Clang-Format
Clang-Format uses the filename (specifically the extension) to determine
the language of the input. When reading the input from standard input,
it uses auto-detection, which tends to be expensive. We can instead
override the language using the --assume-filename option which in my
experiments reduces the runtime by about 20%. Note that the filename in
this case is irrelevant, what matters is the extension.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/941660
Original-Revision: 6aba112ed5b76e168efedd1bfcc5c225885ecd16
GitOrigin-RevId: b5f9b2841a6590982fc229cd89da4286481f433f
Change-Id: Idaca766215e9084a4b085ef97e66ead8ec509bd7