[roll] Roll fuchsia [bt][a2dp] Suspend stream if permit gets taken during setup
The starting of an A2DP stream is a multi-step asynchronous operation.
In some cases during the setup of A2DP audio, HFP audio can be requested
and will take priority by seizing A2DP's streaming permits. Update A2DP
to be resilient to this case by checking for the validity of the permit
before starting the local A2DP stream. If the permit is unavailable,
don't start the local stream and issue a subsequent Suspend to
synchroize the streaming state with the remote peer.

Add a unit test which reproduces this case and fails in the current
implementation.

Original-Bug: b/306661742
Test: fx test bt-a2dp-lib-test

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/939982
Original-Revision: 5328595cfd5f35d4d407de93493b289b894706ea
GitOrigin-RevId: e8e107e6bc88c823a5d3a8285b9fa2db89082b11
Change-Id: I94391fce60d175cada60d6c2eaa73b5a8c295c9a