[roll] Roll fuchsia [build] Wire up product bundles to boot tests
In fxrev.dev/928574, we intended to have each boot_test() define a
product_bundle() under the hood, but it was an oversight on the part of
this reviewer that this did not succeed in actually adding the
product_bundle() into the graph of a boot_test() instance. This change
picks up where that left off and addresses the following:

* the assembly_manifest crate is updated to accept ZBI, VBMeta, and
QEMU kernel metadata of arbitrary names. (The names that matter (e.g.,
zircon-a) only matter at serialization-time into other Assembly
artifacts, but not as Assembly inputs.)

* product_bundle()'s emitted metadata was updated to allow for
user-defined names (as opposed to the hard-coded $product.$board we
reserve for the main product bundle), and also to point to the
JSON-ified version of the product bundle

* make-fuchsia-vol assumed that product_bundles.json only contained one
entry; we naturally break that constraint here and update the tool to
select a product bundle by name.

* Similarly, there was a script
//build/check_product_bundles_manifest.sh that ran during the build to
check that there was at most one product bundle. We update that to
instead assert that product bundle names are unique.

* The metadata file consumed as input to `ffx product create` assumes
that paths within are relative to the file itself. This is different
than the standard assumption that the paths are relative to the build
directory, so we add a simple post-processing step to do the rebasing.

* product_bundle_transfer_manifest() output transfer manifests in a way
that only encoded the target_out_dir and not the target_name of its
targets, leading to collision if more than one instance was instantiated
in the same file.

* The product bundle build API module now includes boot tests.

Original-Bug: b/304279485
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/931527
Original-Revision: d92c6f8aaede062b3b266e4a16f8b5f0bb3019b8
GitOrigin-RevId: 8bb331cd3c65ae2e6ba79c1b86d475e775d1eb38
Change-Id: Ic49491625eb78ae89d0493535804584fac6faaef