[roll] Roll fuchsia [kernel][mp] Clean up CPU offline/online support.
Clean up CPU offline/online support and generalize it to prepare for
suspend/resume, which has significant structural similarities. Instead
of using an ephemeral thread to handle the offline sequence, which can
be tricky to clean up after, expand the responsibilities of the CPU idle
thread to handle idle, offline, and eventually suspend.

The new responsibilities and behavior of the CPU idle/power thread are
handled by IdlePowerThread, which implements the platform-independent
idle loop and the signaling logic for requesting and handling online,
offline, and suspend transitions. Each CPU power thread is persistent
and re-used, avoiding leaks across online -> offline transitions.

Minor updates to the scheduler support scheduling the CPU idle/power
thread when it needs to handle a transition request or maintain the CPU
in a low-power idle state.

The architecture is no longer responsible for implementing the idle
loop, and instead implements an interface for entering idle states based
on latency requirements that is called by the platform-independent idle
loop.

The mp interfaces and tests are cleaned up to remove the now unnecessary
handling of leaked threads.

Original-Bug: 34447
Original-Bug: b/297948013
Test: k ut -r 10000 hotplug; Tracing to verify behavior and performance.
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/923232
Original-Revision: 3897e50fa347b0f53d4cd4766f146805fb7dbd9e
GitOrigin-RevId: 5ba0ada2808fc089cf7ebf370d9c0a6aa6a871a8
Change-Id: I891b2f0401b75e1a63aaa743790de9d491a49968