[roll] Roll fuchsia [testsharder] Fix sharding to check for minimum target duration.
fxrev.dev/934227 introduced logic to be able to decrease the target
duration if it caused shards to exceed the max-shard-size. However, some
long tests can exceed the target duration, so we'd want the target
duration to at least be long enough to run that test once. Otherwise we
get an error because we calculate there needs to be 2 shards of the
smaller duration to be able to run that one test. Thus, we should only
update the target duration if it's at least greater than the minimum
required duration. See go/bbid/8765156804665917201 for an example of how
this failed.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/941858
Original-Revision: 3e2f2883fe4f9eafc633c26d0e588b586974035c
GitOrigin-RevId: 2c18d37cc1d91391d6a910635bf1c4ebf2c5f2c1
Change-Id: Ib3e6e21683d49abc83e20f4fa3ebc14cbe35b398