[roll] Roll fuchsia [fbl] Make fbl::TypeInfo more constant.
fbl::TypeInfo is a helper class which pulls some template tricks in
order to generate a string literal equivalent of a type T's name
(since there is no preprocessor __CLASS__ macro to use).

These tricks involve expanding the __PRETTY_FUNCITON__ macro in a
static member function of TypeInfo<T>, then extracting the value of T
from the expansion into a statically sized name_ buffer which is a
member of a fbl::TypeInfo<T> instance.

To actually fetch the static literal-equivalent of the name, the class
used to say:

```
template <typename T>
class TypeInfo {
 public:
  static const char* Name() {
    static constexpr TypeInfo type_info;
    return type_info.name;
  }

 private:
  // impl goes here.
}
```

The static local provides the instance, which in turn, provides a
compile-time constant location for the storage of the string (all good
so far).  The problem, however, is that the Name method itself is not
flagged as constexpr.

This can cause issues if a user is attempting to initialize another
constexpr instance of a global or static-local class using
fbl::TypeInfo<T>:Name().  While the storage itself is available at
compile time, the call to Name is not constexpr, which means that a
global ctor needs to be generated for the instance that the user is
attempting to initialize (even though, it should not be needed).

It would be great if we could simply slap `constexpr` on the
definition of `TypeInfo::Name`, but that would mean we had declared a
constexpr function with a static-local constexpr variable.  This is
(currently) not allowed, and will not be until C++23.

It would be great if we could avoid this issue by moving the
`TypeInfo` instance to be a static constexpr member of
fbl::TypeInfo<T>, but unfortunately this does not work either.  We
cannot declare a constexpr version of the instance as a static member
of the class because the class has not been fully defined yet, and is
therefore considered to be a non-literal type.

Since neither one of these options are currently available to us, we
work around the issue in a slightly different way (for now).

We can shove the existing TypeInfo implementation into an internal
namespace instead.  Then, we can make internal::TypeInfo's Name
accessors a constexper instanced method, instead of a non-constexpr
static method.

Next, we declare a (much smaller) non-internal version of TypeInfo
which has an instance of internal::TypeInfo as a static constexpr
class member, and a static Name accessor which returns this static
instance's storage pointer.

Noting has really changed, except that our storage has been broken out
into a static member of a different class, instantiated after the
internal version of TypeInfo has been fully declared.  Now the Name
accessor is truly constexpr, and using it to pass a name to the
constructor of another constant class will not trigger a ctor.

In the long run, as soon as we can use C++20 in all of our projects,
we should come back here and re-write all of this in terms of
`consteval/constinit`.  It should change nothing about the current
code generation, but it will provide a compile time guarantee that
everything is, in fact, constant, instead of running the risk that the
compiler is going to start to generate ctors without telling us that
it is doing so.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/940822
Original-Revision: 256f0a042e9f405bcbdbf288fac6fb47d0bbbbac
GitOrigin-RevId: d1d6d8600b531ef0453b4d57afe13454c2e59e98
Change-Id: I016b8e56d94f1a95d43d866ae077c1448cd8c7d7