[roll] Roll fuchsia [testing][realm_proxy] Simple prototype with dictionaries
Demonstrates an "echo test" where the "echo protocol" is backed by a
Receiver served by the Test Realm Factory.

This example also includes code to copy the exposed dictionary to the
test, but it doesn't yet do anything since the API that returns the
exposed dictionary is a stub and will always return an empty dict. This
is tracked in fxbug.dev/303719641.

Original-Bug: b/298100106

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/886692
Original-Revision: fcaf04d0adbb2dbb16b46e3995b22215ee265693
GitOrigin-RevId: 6f885d70e753329052fb6b2d9d16f326690828a2
Change-Id: Ib240d22176537019f82758adf208a4f872e2d948