[roll] Roll fuchsia [netcfg] Add NormalizedMac dyn NameCompositionRule
Extract the normalized mac calculation into a separate function to be
shared by both the new and existing implementation.

Reframe the name retry policy to allow for the NormalizedMac rule to be
modified and generate a potentially unique name with a different offset.
If all offsets are attempted (the value gets to 255), then we will
surface an error that a unique identifier could not be found.

Test: Ensure existing implementation, notably test_get_usb_255, passes.
Added equivalent test for new implementation and test for properly
unique last bytes.

Original-Bug: 135106
Original-Bug: 135087
Original-Bug: b/309010424
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/939913
Original-Revision: bb3e9e797940ec46e69daa1e62fa17e586984309
GitOrigin-RevId: c66e97607e086d420e352a92a1dc5952a5acaf80
Change-Id: Ief42643b366dcf60344153ceb02969cbb483e684