[roll] Roll fuchsia [elfldltl] Introduce DynamicValueCollectionObserver
This observer matches a given tag and adds the associated value to a
provided container. This is useful for callers that need to collect the
values of a given tag type to perform some processing on them at a later
point.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/940436
Original-Revision: 6073d65d0bebf4b277f96df9d5eaffc037da838b
GitOrigin-RevId: bbd964a09618e2ce84f1bf096b4c39d7536fe74f
Change-Id: I4b464d63413f71ee0447a494dfc779da2c4e1c43