[roll] Roll fuchsia [wlancfg] Temporarily ignore unreachable patterns
The upcoming rust toolchain will be able to detect that an open range
covers everything, so the wildcards will be unncessary. This is a
temporary CL to allow us to gracefully update to the latest rust
toolchain. It will be reverted after the toolchain rolls.

Original-Bug: 309552229
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/941677
Original-Revision: 337f30364fded90366d7ed18340362c054543b2f
GitOrigin-RevId: b6128f46edf864393c07f8d055ee7855f72b855b
Change-Id: I8cb83120703c2fcacc1010627bb61c9dd3e2e32b