[roll] Roll fuchsia [ld] Introduce RemoteLoadModule
Introduce the RemoteLoadModule class used in out-of-process dynamic
linking.

This CL defines the RemoteLoadModule data type and introduces the
`Decode` method, which will take ownership of the passed in VMO
and decode it to instantiate the module representation of the binary or
shared object.

Following CLs will implement dependency parsing, relocation, and
loading.

Original-Bug: 134211
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/930549
Original-Revision: 00dd7c4e18150d8aebfa42e570172172186055ab
GitOrigin-RevId: 28f4e55c59d983e16a13a617c6709b1dbee6b8b8
Change-Id: Ibbe6b2a09935effc8aff6b04b28e964c9fa18cb5