[roll] Roll fuchsia [zircon][sched] Fix scheduler counter trace events.
Fix the scheduler counter trace events to use arguments for the values
and the cpu number for the counter id. Use the top level name "Stats" so
that counters are rendered in Perfetto as "Stats:<type>:<cpu>".

This CL corrects a latent bug that was introduced in the ktrace to fxt
format / API migration.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/942272
Original-Revision: 891f418f010a577a32e9877d7825b7e1eabc8fd4
GitOrigin-RevId: af979e5977055ef1ebd73695e579415f0f7087a8
Change-Id: I8cdce4a466ab34e50977a5c46518bce6b27c6f97