[roll] Roll fuchsia [rfc] Amend zx_vmo_transfer_data
After much discussion on https://fxrev.dev/856685 and out of band, it
has been determined that it is not possible to guarantee that the
contents of the source and destination VMO are unchanged on failure.
This update to the RFC reflects that consensus, and instead provides a
new guarantee that the call will not fail so long as certain enumerated
conditions hold.

This amendment also removes some constraints to facilitate Starnix's use
of this syscall.

Original-Bug: 126153
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/939973
Original-Revision: abc5f5fba4275e6561e8ab0870c05bb8905b61fa
GitOrigin-RevId: 99fe138f79f5ba56874bc9ddfb475a41c7277938
Change-Id: Ic4b980d6cbc17de966ab9ad56a3472f3ec4d9c27