[roll] Roll fuchsia [sdk-tools] Allow images in only one bucket.
Currently, this 'gsutil' function will return an error if any of the possible buckets do not have an image that matches the given version string.

I don't believe that is the expected behavior here. If there are two possible GCS buckets but only one of them has it, I (as the caller of this function) would still like to receive the list of available images.

Original-Bug: b/307571441
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/942923
Original-Revision: ca60767f930a2cd5a4befc493472c4af3c8d6926
GitOrigin-RevId: 0b8973ea070ea5eb618a6cd053a34380409b1c34
Change-Id: I4adde706c753be906579e7cc4df410e901c1ad13