[roll] Roll fuchsia [starnix] Update gvisor expectations
https://github.com/google/gvisor/commit/47db4119a2ca7a30d46fbf4a162c01cedb995cfc
was made upstream which fixed a bug in the (now failing on Starnix)
gvisor syscall test `JobControlTest.SetTTY`.

The test previously called `ioctl` with a zero valued FD which led to
`ioctl` returning a non-zero value. This non-zero value was interpreted
as a passing check by the `TEST_PCHECK` macros upstream used.

Now that the test is fixed, it presents itself as a failure on Starnix.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/943592
Original-Revision: 5459ae3cc826751912ce45d06279b0a2e574a68c
GitOrigin-RevId: 811c4e79357375e7b576224e372f4c7d52989fdb
Change-Id: Ic39fd1f3a0bf0e39f2fe686c1e321c2041bc7e7c