[roll] Roll fuchsia [include_bytes_from_working_dir] Add a new proc macro that replaces include_bytes with one that uses cwd-relative paths
The 'include_bytes' macro uses source-file relative paths, which
when used with files generated by the build, will include the
path to the outdir in the env vars used to pass the path to the
compiler.

This is a replacement macro that will instead use paths that
are relative to the cwd, which makes them the same across
build environments, and therefore result in better cache hits
on remote builders.

Original-Bug: 94355
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/942982
Owners-Override: Alice Neels <neelsa@google.com>
Original-Revision: 85503966d0448a218c83982d92e5625de7daa259
GitOrigin-RevId: f586d16df38634a9604a6e123429a3dd660690de
Change-Id: I6787a0311cc8b2e104d8c398e465c28543393840