[roll] Roll fuchsia [netcfg] Add DeviceClass dyn NameCompositionRule
Extract a device's InterfaceType and map it to the existing prefixes to
produce the string that is used in the construction of an interface
name.

Tested: Added tests to make use of the DeviceClass dynamic rule in
combination with other existing rules.

Original-Bug: 135106
Original-Bug: 135087
Original-Bug: b/309010663
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/939355
Original-Revision: 8d1a91d8870fa7d037cce32e6810e837252c3b98
GitOrigin-RevId: 41d59340a5e94316c1e943c84cee7d9306e3d8d3
Change-Id: Iae11a4ceaab7e781a9319ba5dbc3c24dff16108e