[roll] Roll fuchsia [diagnostics][reader] Enable retrying on empty payloads when fully filtered
When a hierarchy is fully filtered, the Archivist returns an empty
payload and an error explaining that it was fully filtered. Clients
using the ArchiveReader API are observing this error when passing a
selector to request data that has not been yet been published. They
expect ArchiveReader to retry until the data they requested is
present. This is a reasonable expectation in tests, which is the
main use case of ArchiveReader.

This change allows to configure the retry strategy that the
ArchiveReader executes which varies among use cases.

Original-Fixed: 136023
Original-Fixed: b/308979621
Original-Bug: b/309698603
Multiply: session-manager-integration-test
Run-All-Tests: true
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/941152
Original-Revision: 0bc6e4aa00626dbd19e89a6995721c8126bdd7e7
GitOrigin-RevId: 2e5fe84442ec8aabc7decf3728808c1c6df9eef6
Change-Id: I8d6a207af6372305f5b993d1e6a625fa905ae941