[roll] Roll fuchsia [lockdep] Add a new metadata only build option.
Add support for a compile-time option to provide access to lock metadata
without enabling lock validation and its memory and runtime overhead.

This change splits up the LockClassState type, which contains both
metadata and validation state, into three types:

1. The base LockClassState type that implements the basic LockClassId
   lookup mechanism.
2. The derived MetadataLockClassState that stores and provides access to
   lock metadata. This type is used when the metadata only is enabled.
3. The derived ValidatorLockClassState that additionally implements the
   state and logic for validating lock dependencies. This type is used
   when lock validation is enabled.

Code that only uses lock metadata may use MetadataLockClass::Get() when
either metadata only or full validation are enabled. Code that depends
on the full validation state may use ValidatorLockClassState::Get() when
full validation is enabled.

Also add the Guard::lock() accessor to the shared variant of Guard. A
previous change added the accessor to the exclusive variant, but missed
adding it to the shared variant.

Also fix a build issue with the tests when lock validation is disabled.
The tests should be able to run and fail when validation is disabled.

Test: Tests pass or faile when validation is enabled or disabled.
      Manually veified that tests fail and no asserts fire whe when
      metadata only mode is enabled and validation is disabled.
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/941655
Original-Revision: 10a6f530316fdc83528418fe84d736d41854269d
GitOrigin-RevId: d3b6ba694a25d0591a00da5d25755cb822d88111
Change-Id: I0b761872b266ac5fe3ae6b67cd05e66a9e6d45bc