[roll] Roll fuchsia [04/nn][vim3] Add SMC and HDMITX MMIO to the vim3-display device.
We add all the resources provided by the vim3-hdmi device to
vim3-display pdev device so that the logic that runs in the
current aml-hdmi driver can be migrated to the amlogic-display
driver.

The definitions of the new resources match those in vim3-hdmi.cc.

Test: no-op expected.
Original-Bug: 134691
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/944297
Original-Revision: d3adc753742ed7ced0670c98146aa64ef6700ef5
GitOrigin-RevId: da68bdf695cc6be593b4b27fb29d487a7b46e8d7
Change-Id: I2d26969ac361971c1606029a37f11da258ac4573