[roll] Roll fuchsia [netstack3] Implement root.Routes
This piggybacks on routes.admin, and the bulk of the change was
factoring existing methods on UserRouteSet to be default methods of a
RouteSet trait.

The main downside to this change is that async-trait has a little bit of
added overhead, which mostly comes in the form of heap allocation.  We
think this is appropriate for control plane APIs, and async traits
in the language proper should be coming pretty soon.

There were two designs I implemented but decided against.  The first was
free functions taking a generic RouteSet parameter.  This worked fine,
but led to a massive diff and weren't the most ergonomic things to use
as opposed to methods (though not by much).

The second was a wrapper struct that was generic over RouteSet.  This
struct was what implemented all of the handler methods.  The upside is
that this removes any of the async overhead, but it made calling those
methods awkward because it meant constructing a struct just for calling
those methods.  Proxying all route set functionality would have meant
implementing an async function (close()) in the RouteSet trait, which
would eliminate any benefit.

The additional Send bound on Responder is required in order to get
async-trait to generate a Future that implements Send.

Now that both netstacks are serving these methods, move the routing from
Netstack2-specific files into the common one.

Original-Fixed: 130863
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/941235
Original-Revision: 64a0ae969cb8922f603d1cabeaa6dde352fb8c7c
GitOrigin-RevId: 8395d8bffb75a3e82911f97afb3e85d9dfc322e3
Change-Id: I539900b48ca558435d9f4c8cd26783ea3c4e3dbf