[roll] Roll fuchsia [kernel][riscv] add support for Zicbom and Zicboz cache instructions
Add kernel support for the Zicbom and Zicboz instructions, needed to
flush the local cache and a fast zero writing option.

Currently enabled all of the time with a cache line size of 64, which is
the standard for RVA22 hardware. QEMU also emulates these instructions,
though most of the time they are a nop.

Original-Bug: 124333

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/906383
Original-Revision: ab57745f31eef356c52271224ae9581a0dfc5ba6
GitOrigin-RevId: 153857d65e930e80f285e33579d60ddaa07d57c4
Change-Id: I39b19d883436f9e98fa00e00d7c88b2d77bcce66