[roll] Roll fuchsia [netstack3] Workaround for UDP-connect to port 0
We've observed real-world instances of applications that rely on
connecting a UDP socket to port 0 to check for presence of IPv4 or
IPv6 internet connections. We need to handle this better with respect
to the socket connected/bound state (as observed via getpeername), but
as a starting point we can check if a route exists to the given
destination and return OK if so.

Original-Bug: 136207
Original-Bug: b/308828371
Test: netstack3_udp_syscall_tests
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/944652
Original-Revision: 99b0491aab4b11e99f1967342031176033837e30
GitOrigin-RevId: 9dabe508e970767ae2a12c141d0c4b7a07b1c9cd
Change-Id: I50b78a850a1da32df40d837a22bd406ad6b66355