[roll] Roll fuchsia [kernel][scheduler] Improve queue state tracing.
Improve queue state tracing event by explicitly signaling whether or
not the QSE's thread is the special idle thread, instead of attempting
to do so by zeroing out the idle thread's reported TID.

Idle threads can experience spinlock contention, just like normal
threads, and the TID used for that report is the actual TID of the
thread, not zero.  Also, the thread's reported name (done at trace
startup) is reported using the real TID as well.  This means that
thread activation events were not being produced for the idle thread's
actual TID (they were being produced for TID 0). Additionally, idle
threads never produced either enqueue or dequeue events (since they
never actually enter or leave the run queue).

Without these events, external parsing tools never synchronize to the
run state of the actual idle threads TID.  This causes the tools to
then ignore lock spin events produced by the idle thread when it
contends on things like the global thread lock, or the global timer
lock.

By changing the queue state events to always report the actual TID of
the thread (and not stuffing in 0 for idle), as well as using one of
the previously unused bits to unambiguously declare a QSE's thread to
be idle or not, we can infer the presence of enqueue and dequeue
events for the idle thread, and properly sync up to its run state,
allowing us to properly measure the amount of lock contention
experienced by the idle threads, instead of simply ignoring it.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/945456
Original-Revision: 61dfc4776af51bbe7f2ee62f30778017dcdeec5c
GitOrigin-RevId: 8674b57aadb20df96cb1a8ee09a58e9dee680e5c
Change-Id: Icbeae513c88e266bf2ce491a9c211a2d8d783faa