[roll] Roll fuchsia [bedrock] Refine Open::downscoped
Open::downscoped always returned a Directory. That turned out
insufficient when writing the end-to-end Router example, because we'd
like to downscope an Open representing the outgoing directory to an Open
representing a protocol node.

Splitting downscoped to separately downscope rights and paths also fixes
a small wrinkle when downscoping to a Directory without rights. That
case is no longer needed.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/943600
Original-Revision: add32235c4f4279076ee3ad643ed1858aad1c76a
GitOrigin-RevId: 1c1c57a34634c676e1ee94333b073818c86d871e
Change-Id: I201142892ecdd75aa2f1b40abf558a549c261b5c