[roll] Roll fuchsia [seriallistener] Write emulator output from seriallistener.
Since botanist already writes emulator serial output to stdout,
seriallistener discarded the serial output it would read so that we
didn't output the serial logs twice. This CL temporarily disables
writing to stdout from botanist and lets seriallistener write the serial
output to stdout instead. This will make it clearer when the boot test
fails to run because it won't show the serial output unless
seriallistener was able to read it. Also, it should help determine
whether the flakes are due to a race between the two processes trying to
write to stdout or not.

Original-Bug: 135386
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/943941
Original-Revision: 4b1d05b4d1aec2e12152b6bacb35fcbd3a8b2540
GitOrigin-RevId: 8f2a2a30ab50136ce4b635ae65914b8501107f56
Change-Id: I7e3ff0dea093a0ec5912f25d8f0120897c5c542a