[roll] Roll fuchsia [designware-hdmi] Rename "hdmi_dw" namespace to "designware_hdmi".
"dw" is not a common abbreviation and it's not obvious that it
stands for "DesignWare". This change renames the namespace from
"hdmi_dw" to "designware_hdmi" so that it uses the full name of
the IP core name, and it follows the existing pattern of other
driver names ("manufacturer-product", for example, "amlogic-display"
and "intel-i915").

Test: no binary change.
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/945632
Original-Revision: 85add6eec26177b997447ae9ccb32cf10fcb0ac4
GitOrigin-RevId: bb20ec8aa18af01855e8e141ddf568d10efc0da3
Change-Id: Iad026b2cfd7ef77b290ab7eb4d0995d19087f258