[roll] Roll fuchsia [compat] Name DFv1 driver VMOs
The driver host names each of the DFv2 driver VMOs as they are
created. Do the same in the compat driver, so that it is easier
to read memory usage in a driver host.

In order to name the vmo, it had to be opened with kPrivateClone,
which is what DFv2 driver vmos are opened with.

`ffx memory profile` was taken before and after this change on
QEMU to check that memory usage was the same.

Original-Bug: 309972262
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/944694
Original-Revision: a89b403b1465efc8e5a719b457d12c23147f8f14
GitOrigin-RevId: 6a03b3ca1da4aa8bd0eb2671a533866070191beb
Change-Id: I1b79e08052ac17c8ed3809c502ceae422b4cd3c4