[roll] Roll fuchsia [component_manager] Availability is required
Remove the Option from here and update the default to
Availability::Required. This is the de-facto default.

We can see that the Option is not being used by the fact
that it is always unwrapped with 'expected'. Updating
this removes the chances of Component Manager panicking
at runtime on a bad expected.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/946112
Original-Revision: 2582994649babce0cb6f6472c49c3532f70b544e
GitOrigin-RevId: bd622ee2a1a507f6f4c801667982517846862933
Change-Id: I7bccf8b20e305969b60a1a717cdb40ae6c777a37