[roll] Roll fuchsia [testrunner] Use sshTester's EnsureSinks() if ffx test not enabled.
The FFXTester may be used if ffx is enabled but `ffx test` is only
enabled at experiment level 2. Thus, the way to collect sinks from the
ffx test outputs should also only be enabled when `ffx test` is used.
The coverage builders broke when fxrev.dev/914436 forwarded the ffx
flag to testrunner, causing it to use the FFXTester with experiment
level 1 which means that it was NOT using `ffx test`.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/945979
Original-Revision: 84863f63c7364513a6e49617037eebd25850f42d
GitOrigin-RevId: 0646c00e316c0feb55ac94822b680d66702ae34e
Change-Id: Ia5c6c0023f7a37838e69e60f239182b274be6030