[roll] Roll fuchsia [kernel][vm] Fix unified_accessed unittest
Fixes the vmaspace_unified_accessed_test to work on ARM. Note that this
change does not enable the unit test on ARM yet, it just modifies the
test to take into account the fact that ARM sets the accessed flag on
Map operations. This was not a problem in the past because x86, which is
the only architecture on which this test runs currently, does not set
the accessed flag on map operations.

Original-Bug: 132980
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/945269
Original-Revision: f6268dd07d1d26154473c486e139696917b16aee
GitOrigin-RevId: a46dd668c72ae43770b7167599257b3c7ed9b971
Change-Id: I4bb09c40a5120425a67b46d6900050bc20366add