[roll] Roll fuchsia [build][ldso] Avoid LTO codegen bug with inline ASM
Clang has a bug whereby module level inline ASM requiring target
features may result in codegen errors that will not cause a link to
fail, but may corrupt the symbol table (e.g., by omitting the function).
Note that this only affects module level ASM, and not all instances of
inline assembly, so the number of places this affects in Fuchsia  is
relatively small. Thus far, it only appears to affect riscv64 in our
codebase, but the bug affects all architectures, and needs to be fixed
in upstream LLVM.

Specifically, we see this causing an issue in dynlink.c. The proposed
workarounds for now are to disable LTO or to move the ASM into assembly
files. This patch adopts the later approach instead of disabling LTO.

More details on the codegen issue can be found on
https://github.com/llvm/llvm-project/issues/67698

Original-Bug: 136220
Test: fx set bringup.riscv64 --variant=lto --with //bundles/boot_tests --auto-dir
Test: fx build
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/945114
Original-Revision: 854e9869ba8744b903524a8aab500881dec97420
GitOrigin-RevId: 4c3db69d27b412b86bbc879ba2219725bfcc25de
Change-Id: If97b58732b697d64f4a92c2553cf15ab5b87f468