[roll] Roll infra/recipes [test_owners] Avoid infinite recursion
Previously we'd only exit the loop upon hitting the root of the checkout
if the first-level parent directory of the test in question didn't
contain an OWNERS file. So for example if a test was declared in
`//src/foo/bar/baz/BUILD.gn` and the only OWNERS file in that directory
chain was `//src/OWNERS`, we would keep traversing up the directory
structure past `//` and end up looping forever upon reaching the
filesystem root.

Now it's no longer possible to escape the repository root and enter an
infinite loop.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/infra/recipes/+/946712
Original-Revision: 5fdd98ac7b46dd11fbc6f7efa377eb6ea1f6a88e
GitOrigin-RevId: 1cf8922468c2ec10523fcd008373473fe7e56542
Change-Id: Ic53bcbcc9b7e63e195ab44e4952687355d6d12d3