[roll] Roll fuchsia [audio] Correctly translate driver formats
With this CL, the ADR expressly fails to load an audio
driver, if that driver declares that it supports a driver
audio format (in fuchsia_hardware_audio) that cannot be
translated into a valid framework audio format
(fuchsia_audio).

Previously, this would have succeeded, then later failed in
a way that was misunderstood by the developer.

This CL also adds a very significant chunk of unittesting
of the external validation functions. The functions that
are not yet unittested are labelled with TODO placeholders
in the unittest sources. I will continue to add to these,
and they may be complete by the time this CL's
dependencies have landed.

Original-Bug: b/307787998

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/945562
Original-Revision: 04818d3997ed07404a26b7e4d1c160fe5e7ba38f
GitOrigin-RevId: 0cf821c25fbcca53ec0b318a69deaf3978df93dd
Change-Id: I78cfccc424f9fa55e5ececa1eaf6ed0743429d9e