[roll] Roll fuchsia [sdk/fuchsia.test.manager] reorder definitions
This CL reorders the definitions in test_manager.fidl so they are in
top-down rather than bottom-up order. It makes no other changes.

In a subsequent CL, the same file will be updated for conformance to
rubrics and the FIDL style guide. The updated definitions will be in a
similar order, and this CL provides a better diff base for the new
definitions.

Test: no behavior changes

Original-Bug: b/309995479
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/946744
Original-Revision: 88b03f2334d0334b5bb72f2bd6294bb58bd01281
GitOrigin-RevId: 989d8aa60d4ccac05a8a00c19906db1612243826
Change-Id: If9a38ff7bdd54def0ccbdfaa9a527ecc2f81aeec