[roll] Roll fuchsia [fuchsia-controller] Update IR path finding.
Uses the default config env to look for the IR path. Also adds a build
declaration to point to the default IR path within the config.

Will now not rely on using the build dir files or looking for specific
file markers to find the IR.

Original-Bug: b/309489520

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/942034
Original-Revision: 28188be039860c8327bda5237812d19351676273
GitOrigin-RevId: 7e529a0f32084efd626e26bbfcb972b5035f4264
Change-Id: Ice82245f169810bb4092cf5cc8a73a8ce1dba306