[roll] Roll fuchsia [performance] Automatically generate the most simple benchmark Python drivers
Instead of requiring benchmark authors to write Python host-side code
to drive their benchmarks, we can ask them to call a GN rule that
runs a common parametrized test for them.

In my opinion this improves the DX of writing a benchmark, given that
users can focus on writing their target-side code and then just bring
the benchmark in the build so that it gets automatically executed and
the results published (as long as the expected metrics accept them,
which continue to require Performance+2).

This change implements the minimal functionality, we can expand as we
run into the need of more functionality while we migrate tests to the
new infrastructure.

Original-Fixed: b/308500038
Cq-Include-Trybots: luci.fuchsia.try:terminal.vim3-release
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/938212
Original-Revision: 333fecfa86425a88d874307010a3a572d039f10e
GitOrigin-RevId: 132dc159e572a099a184906ec983c98dd72128dc
Change-Id: Idf44fe89c4d5cf776bb4e09300e25473daa32f72