[roll] Roll fuchsia [rfc_appscript] Don't revert status.
Addresses an edge case in the automation: if an RFC author moves their
CL back to "work in progress", it'll currently kick the RFC back to
"socialization".  That may seem reasonable at first glance, but I could
imagine someone getting the notification that an RFC is "on hold" and
then marking their CL as WIP, sending it back to Socialization. Then, 4
weeks later it'd get set back to "On hold" again, generating another
annoying email.

So yeah, this fixes that.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/947272
Original-Revision: e3319bc2a0aefdac216936816fd547487d7e8252
GitOrigin-RevId: eaac2cb4ed9eda50fd45a6a52c819d6d360e5be9
Change-Id: I4ee5abdd82936e28d8933c9d6586a60438788e5d