[roll] Roll fuchsia [driver][async][sdk] Add async_helpers lib
Adds a new library that can be used to help with simplifying async
cancellations using a cancellation token and work tracker. This is
useful as we have a lot of async code in drivers and currently they
have a lot of client lifetime extension mechanisms that don't allow
for cancellation.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/945457
Original-Revision: ef48c94d7365e0fa188709d159f7a111191b10a9
GitOrigin-RevId: a208a4bc8108242a8842431663a5f4435f329f86
Change-Id: Ib4951aa0e40b296353824cb24f7833368f35d5a7