[roll] Roll fuchsia [fuchsia-url] Support fuchsia-builtin:// scheme
The fuchsia-builtin:// URL scheme is a simple URL scheme that
identifies a component declaration contained within component_manager.
This CL adds support for parsing those URL to fuchsia-url, such that
we can reuse some of the validation logic.

At a high level it could be thought of as fuchsia-boot:// but where the
path must be "/". It is intended to resolve components such as the ELF
runner.

Original-Bug: b/305862055
Original-Bug: b/309471663
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/944824
Original-Revision: 2640590b896084587d7154003437cc93906c9a81
GitOrigin-RevId: 5f084e7d5dcec77d4c8a2b2afed458e1e828c11f
Change-Id: I2b1c41ca08b05f232706f7328bb2ef993fa21036