[roll] Roll fuchsia [netcfg] Name interfaces based on multiple rules
Determine whether an interface matches ALL of the matchers prior to
generating a name based on that selector's naming scheme.

Create a helper fn to check whether all matchers for a given
`NamingRule` match the provided `DeviceInfo`. When the BusType cannot be
parsed from the topological path, assume there was not a match instead
of surfacing the error. The error for not finding a BusType will be
surfaced during the naming portion if that dynamic rule is used.

Original-Bug: 135097
Original-Bug: 135087
Original-Bug: b/309010485
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/943936
Original-Revision: 320d8ee7aa25003942ad9a7952716c488ce40da2
GitOrigin-RevId: 84694c9453407af544258dd19428f5e2a2ee9687
Change-Id: Ica910b9ede2ce0766f2966e6431e60021e4f86fd