[roll] Roll fuchsia [bedrock] Replace Sandbox with Dict
The Sandbox type is a component_manager-specific version of Dict
that provides methods to get and insert "CapabilityDict" entries,
which contain a Sender/Receiver pair and availability.

To reduce confusion between the sandbox library, and the Dict type,
Sandbox is now a plain dict, and DictExt is an extension trait
that provides methods that operate on CapabilityDict entries.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/941580
Original-Revision: 5c1c26d7aea213f27fedf9c9f6ef291716267709
GitOrigin-RevId: 6fcd4ad714492266d20bd09d09a39a90d8f83f6b
Change-Id: I860f7900022ddb03254895e7a8873b6ab5e2edd4