[roll] Roll fuchsia [ffx] Move fuchsia.developer.ffx out of /sdk
All of this is only used by ffx to communicate with itself.

We leave a couple of the daemon methods in /sdk in the new
fuchsia.developer.ffxdaemonlifecycle protocol so that we can put them in
the SDK later, and thus maintain daemon compatibility for that small
portion of the interface.

Original-Bug: b/291987067
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/937217
Original-Revision: b07a167bc12c3ac87a2277498019a7b9ad3d677c
GitOrigin-RevId: f5a84e0682b6627138386251525c0d251b23b73a
Change-Id: I853da094835a052fa67cf498796ebeb838634fc6