[roll] Roll fuchsia [designware-hdmi] Split the interface type and the implementation type.
In order to test aml-hdmi independently, previously we created a test
double for HdmiDw (FakeHdmiDw, which is actually a mock). Some of the
methods that may touch the HDMITX MMIO region were defined as virtual
methods being overridden in the FakeHdmiDw class.

That mock class inherited parts of the HdmiDw functions (EdidTransfer /
InitHw) and mocks the other methods, which can be very confusing for
the test developers. The existing FakeHdmiDw was also highly coupled
with the AmlHdmiTest fixture, making tests hard to change.

This change splits the HdmiDw to a pure abstract base class
(HdmiTransmitterController) and an implemnentation class
(HdmiTransmitterControllerImpl). Test doubles inherit the pure
interface instead of the implementation class.

This change also renames FakeHdmiDw to
MockHdmiTransmitterController which matches its behavior. It also
decouples the mock class from the test class.

Test: aml-hdmi-test and hdmi20-transmitter-controller-impl-test passed.
Original-Bug: 136258
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/945633
Original-Revision: f99a7ec7812ab78c9f789929fff40ce5588172ad
GitOrigin-RevId: e3c1526932952e11f0b6a292124fd9514bd8bf5c
Change-Id: I531d6d7b1b85a688505d4c69ce835ca1678ff016