[roll] Roll fuchsia [fbl] Add utility for load/stores to unaligned addresses
Loading or storing a value to/from an unaligned pointer such as an
offset into a char array or a packed struct is tricky to do correctly in
C++ and is needed in various places in kernel and userspace. This adds
a pair of fbl utilities for this that invoke the arcane rituals.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/947540
Original-Revision: 4021547abc6e7eb83d5047f2a3c7332261ca07be
GitOrigin-RevId: 918069dd54a70014c0e1e7176a57ab74d4ac5417
Change-Id: I402083d4cce669855e96442bd6d0ed282bd9e31f