[roll] Roll fuchsia [audio][drivers] Remove deprecated external_delay
Previously, an audio driver's 'external_delay' was moved
from the static RingBufferProperties to a dynamic DelayInfo,
returned by the `WatchDelayInfo` hanging-get. Although the
external_delay value in RingBufferProperties was marked as
deprecated, a number of drivers still returned it.

With this CL, no in-tree audio drivers set a value for
RingBufferProperties.external_delay. This is a prerequisite
for actually removing the field from the FIDL table at some
point in the future.

Original-Bug: fxb/116898

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/947368
Original-Revision: 5966ea710b89bfd9430e458a88937fca55a486ab
GitOrigin-RevId: 3dc532f562766d5fdf9f23a03072904e0a8e7b05
Change-Id: I7b6aa04ea5613d8516334a31f94b1a569b2ea532