[roll] Roll fuchsia [component_manager] Factor out shared availability transformation
This is a useful primitive that appears both in production and in
playgrounds. In doing so I also added tests since there's a deficiency
in unit tests in some capability types.

This answers the question of "should we check relative paths" in
new_forwarding_router in the negative.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/947546
Original-Revision: 769b5011f8494538565f289dbce9fbc6f2eaba1f
GitOrigin-RevId: 0b98f4ccc430980c9e440cf4029c0a1388f6ff9e
Change-Id: If8c446a948d773ed234abfe3be85249d035f1d01