[roll] Roll fuchsia [component] Migrate a bit of integration tests to use an ELF runner component
We have quite a few integration tests that spin up a different root
realm constructed from scratch. Because "elf" will eventually disappear
from the root environment (see fxrev.dev/940154), we'll need to find a
way to keep running these tests which still need an ELF runner.

This CL adds two shards that test root realms may include, to add an ELF
runner component back to the topology. It is quite similar to how the
ELF runner would be added in production, so should give us good
coverage.

An alternative that was attempted that didn't pan out was to add a
component_manager config that will cause component_manager to register a
runner named "elf" in the root realm that delegates to a namespace
ComponentRunner FIDL, similar to the builtin realm builder runner.
This didn't work because we do a KOID lookup on the controller server
endpoint to determine the moniker. When running a nested component
manager, the outer component manager does not know about the controller
server endpoint minted in the nested component manager.

If this CL looks good, a subsequence CL would migrate most other
integration tests to this pattern.

Original-Bug: 305862055
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/947366
Original-Revision: 0e7c81a490db1f097c7fedc5871821a872f2e107
GitOrigin-RevId: 555e528f63c3dead0b698308f2b4f6d792292a9a
Change-Id: I3ff83397340623efad22770147c217de235caddb