[roll] Roll fuchsia [HoneyDew]Re-enable Fastboot functional test in infra
fastboot test was failing sometimes because of a known platform issue causing the
device unhealthy and thus causing all subsequent tests to fails. So we disabled runnng this test.

Now that FuchsiaBaseTest has device recovery, we can reneable the
fastboot test

Re-enabled the test and ran 16 times in patchset#3 and was able to reproduce the
fastboot failure and made sure recovery logic kicks in and recovers the
device and subsequent test are now running.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/946618
Original-Revision: 394cf15afaa31386452f4c8e162aa0137ed9c6fa
GitOrigin-RevId: c747cf3ad04b51d2c54c6d214d6d6043c6d9ade4
Change-Id: I76ce9da121680e931976c79b678d0b594f1d5f6c