[roll] Roll fuchsia [bt][hfp-hf] Add delay handling search results.
Previously, we would immediately attempt to connect to a peer on a
search result.  However, this could race with the peer attempting to
connect to us.  Since it's more likely a peer will want to connect
first, this CL adds a delay to allow this before attempting to connect
to the peer.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/864062
Original-Revision: 5b0b4398cb040e9f4a969e774911afcd05c20ea5
GitOrigin-RevId: 745578bc322d871f8d48dd2c783ecb40bc6b687a
Change-Id: I6e71b1f082f1fcd968c1990316b491c3cbeb98e6