[roll] Roll fuchsia [debugger] Add Linux helper routines.
Adds a utility that will convert a block memory writes to a series of
aligned 64-bit writes. This is expressed in a generic way so that it can
be unit tested. It will be used for ptrace memory writing which can only
write 64-bit aligned memory words.

Adds a BufferedStream implementation for a Posix bidirectional pipe.
This will be used to channel debug_ipc over a pipe.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/945419
Original-Revision: 18d59865c5774adda722a0d07c7fd257d12916d9
GitOrigin-RevId: c0b0f46371c36694623d739dd378bc71b3a206a0
Change-Id: Ib635d59f89b7fdd771093bd8f9b9dce8eabda917