[roll] Roll fuchsia [netstack] Remove GetDnsServerWatcher
There are no longer any users of the method.

Normally, we'd want to split off the Netstack2 and Netstack3 changes
from from the removal of the method, but that leads to compiler errors
due to unhandled messages.

Removed the TODO on dnsServerWatcherCollection.Bind() without removing
the method because it's load-bearing in tests.

Original-Fixed: 134162
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/937198
Original-Revision: dbb014645e1dafc3c434cbec42b71213f978b4ea
GitOrigin-RevId: 80151b791f7138c77c6611c536eaa4e1bcd642b5
Change-Id: Ib3b9a864f57e08212cc8e14e9824ba291af89568