[roll] Roll fuchsia [power-manager] Prevent busy-loop on closed activity provider
If the watch state stream closes, watch_activity attempts to reconnect
by calling fuchsia.ui.activity/Provider.WatchState. However, this FIDL
method is one-way, so succeeds even if the underlying channel is
closed. This results in an infinite loop in watch_activity which spams
reconnect logs and also, I believe, prevents the component's single-
threaded async executor from making progress on any other tasks (
because the only .await in the loop is on the stream which I think
is guaranteed to be closed at that point and so poll_next always
returns Ready(None) which I'm guessing prevents the executor from
polling any other tasks).

This change checks if the fuchsia.ui.activity/Provider connection is
closed before calling watch_state, which should prevent the infinite
loop and hopefully prevent some flakes in the OTA tests.

Later work should determine why the Provider connection is closed
and possibly move the closed check earlier in the process or make it
blocking and possibly add reconnect logic to the Provider connection
itself, given that the code is interested in reconnecting to
WatchState.

Original-Bug: b/311783026
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/948626
Original-Revision: 2113a962e067b9b8f24d9411e57c68128bd69a43
GitOrigin-RevId: 734a78553aa107baa37905df92fd4243bfa8b261
Change-Id: I584345602a419abf8d402833a3fc697b1ae922d5