[roll] Roll fuchsia [a11y] Backdate fuchsia:accessibility::semantics::EnabledState API to v15
This change is using the "escape hatch" described in
go/Smart-Components-Stable-APIs-Escape-Hatch. It is being used because
the use case for b/261482081 falls under use case 2 of "Targeting
In-Development APIs" in go/sd-petals-targeting-stable-apis.

This is a newly added FIDL API that was added in fxr/941818, with no
existing usages.

Original-Bug: b/261482081
Original-Bug: b/311150772
Test: n/a
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/946310
Original-Revision: 050bf35685258e6f3e73afc2c5b25d878d65a449
GitOrigin-RevId: f4a518a7a82e9b9a6ee30cd1b2cf4f37f5cef7c7
Change-Id: I5ace59c74c0b57f485b21673210a5a7ed0cafd85