[roll] Roll fuchsia [netcfg] Correctly classify AlreadyExists error
When an interface name is generated and it is the same name that already
exists, it was being classified as a Fatal error for the
`handle_device_instance` workflow in netcfg and crashing the component.
This should instead be treated the same way as if an interface is
rejected from the Netstack with an AlreadyExists error, where temporary
names are attempted if the first one is unsuccessful.

Change the NameGenerationError::AlreadyExists to map to the
devices::AlreadyExists error instead of a fatal error so that we can try
alternative names instead of crashing.

Original-Bug: 136406
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/947612
Original-Revision: afe151523194dfb4ef755aca02bd34ed545f8421
GitOrigin-RevId: 2bab35bc597efabc6c688168c50a1a1f2506f9fb
Change-Id: I435b2852374379ca9aad8d0eb135c54b6d2700eb