[roll] Roll fuchsia Reland "[netcfg] Add default dyn NameCompositionRule"
This is a reland of commit e17b8d121c455b6b459304a46c9ed23120659cae

This CL made use of the `generate_name` function which produced
a NameGenerationError::AlreadyExistsError when the name generated
was the same as one that already existed. In the caller, this
error was mapped to a Fatal error, which led netcfg to crash.

The fix in fxr/947612 should amend the cause of the crashes in CQ.

Original change's description:
> [netcfg] Add default dyn NameCompositionRule
>
> Write the current naming structure for USB and PCI/SDIO as equivalent
> to specifying the `Default` dynamic rule.
>
> Refactored existing implementation to use the default naming rule
> instead. This required making some sample values, but enforces that
> the new implementation reaches parity with the existing.
>
> Tested: Added tests to make use of the default dynamic rule for the
> USB and PCI/SDIO scenarios.
>
> Original-Bug: 135106
> Original-Bug: 135087
> Original-Bug: b/309010663
> Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/941654

Original-Bug: 135106
Original-Bug: 135087
Original-Bug: b/309010663
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/947816
Original-Revision: e1dcfe36784c05c1c6b7812a6fe0a88daa9b7a82
GitOrigin-RevId: e03cbf7cc66ea0bfdad4448b60d83c7141d1c6dc
Change-Id: Ia0d72fdd8131a1aa33fca354ffe2027199a71ceb