[roll] Roll fuchsia Reland "[netcfg] Name interfaces based on multiple rules"
This is a reland of commit 320d8ee7aa25003942ad9a7952716c488ce40da2

This CL's upstream made use of the `generate_name` function
which produces a NameGenerationError::AlreadyExistsError when
the name generated was the same as one that already existed.
In the caller, this error was mapped to a Fatal error, which led
netcfg to crash.

The fix in fxr/947612 should amend the cause of the crashes in CQ.

Original change's description:
> [netcfg] Name interfaces based on multiple rules
>
> Determine whether an interface matches ALL of the matchers prior to
> generating a name based on that selector's naming scheme.
>
> Create a helper fn to check whether all matchers for a given
> `NamingRule` match the provided `DeviceInfo`. When the BusType cannot be
> parsed from the topological path, assume there was not a match instead
> of surfacing the error. The error for not finding a BusType will be
> surfaced during the naming portion if that dynamic rule is used.
>
> Original-Bug: 135097
> Original-Bug: 135087
> Original-Bug: b/309010485
> Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/943936

Original-Bug: 135097
Original-Bug: 135087
Original-Bug: b/309010485
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/947893
Original-Revision: 02a4caf64ae95302933bd1591965dd66a8b4ffd8
GitOrigin-RevId: c14672823b452854bdacd0adc03b728742f629ac
Change-Id: Icf447923b40d05380c5dccd3cc93d715e8c5d099