[roll] Roll fuchsia [netcfg] Enable non-default interface names
Prior to this change, there was not enough information for interfaces to
call all other `NameCompositionRule`s and produce a functional name. For
example, a device that was named with a MAC persistent identifier would
cause a crash if the configuration specified something such as the
BusPath because there would be no topological path present to parse.

Pass through the fields directly when the device is discovered, so
that the name can be produced using all of the available information.

Create a struct that mimics the `DeviceInfo` struct and uses references
instead of owned structs. Also, it confirms the MAC is present, since
devices without MAC are not supported by netcfg.

Refactor some tests to pass in `DeviceInfoRef` instead of a
`PersistentIdentifier`.

Original-Bug: 135109
Original-Bug: 135087
Original-Bug: b/309010488

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/945355
Original-Revision: 20f165d774f12f55bf41be7161fd7774295f9af8
GitOrigin-RevId: 78e44adf1f9c8c48e42e1cc188f06c4e1495185c
Change-Id: I51d6127b937c8260b7d620f850654d3405f9f986