[roll] Roll fuchsia [zxdb] Limit support for "run" to Linux
This command runs the specified binary in the debugger's namespace,
which is very unlikely to be what the user wants. Instead, users are
much more like to want to use the `attach` command to attach to an
existing process.

This CL limits support for "run" to Linux targets. Also, remove the
incorrect warning when using "run" on Linux.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/939364
Original-Revision: 32369ab4e31d7a79ae9780eb45c80fc0240881c7
GitOrigin-RevId: a68b70b70e2e38b76208bd4681684971554ee4f1
Change-Id: I4f6df1feaa6029b3b3c1edcfbb4623de363cbc86