[roll] Roll fuchsia [superproject] Roll third_party/Vulkan-Tools [logs] Depend explicitly on syslog library
Previously, this was indirectly depending on the syslog C++ library.
This used to work when syslog was non-hermetic due to not being
exported to the SDK. When we export to the SDK, it will need to be
made hermetic which breaks symbol resolution.

Original-Bug: b/300157652
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/third_party/Vulkan-Tools/+/949072
Original-Revision: 9f0f40b6429aed1a8371142bfc7e31adf3d07c30
GitOrigin-RevId: 67626999e4b4e78d1cd77f7913329e7dc5be976b
Roller-URL: https://ci.chromium.org/b/8763880585426129777
CQ-Do-Not-Cancel-Tryjobs: true
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/949516
Original-Revision: d0e157205392c85796629d0683cc249fe6b9f928
Change-Id: Ifaac916063710998aa3afde96659818ca51b7435