[roll] Roll fuchsia [ffx] Add ssh ConnectTimeout option
Sometimes it appears an emulator can listen on a port, and then
never forward the port to the guest O/S. At least that might
explain why ssh connections to the emulator sometimes just
stall when we try to establish a host-pipe immediately after
the emulator starts. By adding a timeout, we have a chance
to re-try the ssh connection.

This may not work, because the ffx emu's `is_active()`` function checks
for a host-pipe error, not waiting for a retry. But we might at least
get better diagnostics this way, even if it doesn't work.

The timeout has to be chosen somewhat judiciously -- too short and we
_always_ hit the above error.

Original-Bug: b/301312339
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/948621
Original-Revision: 16ae63db2dcb1cc9ff3adc6db3f56e0f39686282
GitOrigin-RevId: aa2370ad0a83ae9c2743a2ed1426393cc25355f6
Change-Id: I9104bc972f9d710b979a9b8391976524d53cda2b