[roll] Roll fuchsia [build] Extend product bundle support for boot tests with EFI disks
This aspect of boot test support was missing.

If the boot test specifies an EFI disk image, then this should be
regarded as the contents of a bootloader partition - and in this case we
synthesize a basic config file; otherwise, we default to the
board-specific config.

Original-Bug: b/304820366
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/947907
Original-Revision: 3d9d5773f44a18cc28948ee34c550c804ad53ed3
GitOrigin-RevId: 1958c3310414547e9abe8566d9d402b7def4e09f
Change-Id: Ia34d630721a3da6006790aab4e79b7b939009f66