[roll] Roll fuchsia [debugger] Allow local Linux debugging in zxdb.
Enables the functional but incomplete version of zxdb. It is enabled by
passing the --local switch.

This uses a built-in debug_agent that is automatically forked on
startup. There is not currently a separate Linux implementation of
debug_agent for remote debugging but this is easily possibly by setting
up the right socket connections.

Restructures some of the connection initiation code to share code
between the various implementations (local, remote, minidump). This also
hooks up more state like the architecture for the minidump "agent".

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/412296
Original-Revision: 6bc9f4f4aeafa3ce6598304fd3e85c001c55468e
GitOrigin-RevId: a0cc1fcdf35779bdc34edf43e8bd5fec9cd37416
Change-Id: I14e3846a0f709f30434611989be20379c0660759