[roll] Roll fuchsia [netcfg] Add InterfaceName MatchingRule
Intake a pattern from the provided configuration and determine if that
matches against the interface's name. This matching rule only
available to be used in the matchers for the provisioning policy - the
configuration will not allow this option in the naming scheme for
naming rules.

Tested: Added tests for matching an interface name with various glob
expressions.

Original-Bug: 135108
Original-Bug: 135087
Original-Bug: b/309011118
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/943337
Original-Revision: 403815561d28debe1bf5ccd3556966a5ea40c8b6
GitOrigin-RevId: 8cbe4c18470ce48dda428762434537d7e0b293c5
Change-Id: Ie12ba0724dd6ac7d2df9fe78bdf66dcc150110bc