[roll] Roll fuchsia [debugger] Don't require slash for local symbols.
Previously, loading local symbols required that the module name begins
with a slash. This was done when the local fallback was kind of a hack.

Now that we have a strong notion of whether the connection is to the
local system, we don't need this extra gate before checking for local
symbols. This allows debugging Linux processes without specifying an
absolute path to "run".

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/950014
Original-Revision: 9ef645760f4f05f99a6996479fc7b3a9b050d34f
GitOrigin-RevId: c42109558ad11a03feeb41b885f51a29ea48ef69
Change-Id: I58f81220483748886247689875e2bd23fbe32c24