[roll] Roll fuchsia [windowed-stat] Update CombinedWindowedStats implementation
CombinedWindowedStats is now time-aware and automatically slides the
windows if the current time does not fall into existing windows. This
automated operation occurs when the client logs into the
CombinedWindowedStats or dumps it into Inspect.

In this implementation, the bounds of each window are no longer at the
timestamps when windows are slidden. Instead, the bounds of the windows
are always fixed (or prealigned) at the timestamps that divide the
granularity. For example, for the `1m` bucket, the bounds are at the
0th, 60th, 120th second mark, and so on.

CombinedWindowedStats is now renamed to TimeSeries.

Original-Bug: 135511
Test:
1. Verified timestamp is logged
2. Updated unit tests

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/934810
Original-Revision: ef7e38bbeacd314fca0d1d346efaa44ec53e5f6f
GitOrigin-RevId: 0aaa15a7ff8f7803baa145c3d828028dfdc10ea9
Change-Id: I517bb7fa90859e8645685ad4913dab983b1ac1ab