[roll] Roll fuchsia [bt][hfp-hf] Make SCLI take one AT resp at a time
Currently procedures take a vector of AT responses.  However, this
doesn't work, as we can't guarantee that the underlying RFCOMM channel
will produce all the required AT responses in a single data transfer--
in fact we expect it to not do so.  This means we need to correlate
which AT responses go together into one vector to make the pattern
matching in the procedure transition methods work correctly, which is
hard to do outside the procedure.

To fix this, I am changing the procedure trait transition method to take
a single AT response.  This requires changing all the procedures that
pattern match against multiple AT commands, before changing the
transition method type.

Fortunately, the only currently implemented procedure that needs to be
changed is SLCI.  Several states in the procedure state machine for SLCI
expect an incoming response followed by an OK.  I've broken these
states into two states, the original and another for waiting for the OK.

I've updated the tests to match.

Original-Bug: fxb/128997
Test: fx test bt-hfp-hands-free-tests

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/878814
Original-Revision: 3f56caa31781590593fe0e42ef72dacd364ff090
GitOrigin-RevId: a35649bf2ff3432d353a5513809ad2f8f46f1d75
Change-Id: I61eb97b9520a422e6ea10f1e4d775f9c9ebc9b29