[roll] Roll fuchsia [swd-e2e-tests] Non-max ota tests should just add a single file
Previously, the non-max-size OTA tests worked by actually treating it as
a max-size OTA test that was just one block larger. However, this didn't
account for adding the reference to the `meta.far` could also push it
larger than a single block. This could cause the test to flake.

This change gets rid of trying to fit the non-max-size OTA tests within
a bounded size, and instead just adds a single file without any
checks.

To help simplify debugging, the generated files are now written
with their merkle hash.

Finally, this adds a safety check to make sure that `packages.json` and
the `system_image/0` are kept in sync, to both make sure the OTA tests
don't make a mistake when generating packages, and as a notification for
us if we ever accidentally make the two lists out of sync.

The only thing we expect to be different between the two at the moment
is the `subpackage_blobs/0` meta.far merkle, which is used to make the
subpackage blobs are available on the system. Technically we don't need
this anymore since F11, but we need to keep this logic around until we
stop producing the `subpackage_blobs/0` packages.

Original-Bug: http://b/312027540
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/949500
Original-Revision: c92b4ece7b2d195aa4f382a87e021da6157a19a5
GitOrigin-RevId: 3d1a83d0752f17a29ccc17f3ea2c8ab68be2a266
Change-Id: Ic657b9bb21058250e0b5294d1721a529489a881b