[roll] Roll fuchsia [fxtest] Implement --updateifinbase
This flag was present but did not work properly due to some issues with
the update-if-in-base script. The new implementation properly parses the
base package lists and if necessary will build the update package and
perform an OTA.

During local runs the OTA will fail, so the result is printing a warning
in that case.

Original-Fixed: b/295340410
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/950204
Original-Revision: 552daac15ad2d1e3459157186f7371a1f5a68a38
GitOrigin-RevId: 2037b90490dc61134cbb38e02c4d16d904091986
Change-Id: I5c1e037f64b04190dec1fa32b921627a0ae1f179