[roll] Roll fuchsia [swd-e2e-tests] Support the new log listener
(This is a re-attempt at http://fxrev.dev/948776).

This can use the new or old log listener to check for unknown
firmware error messages.

We read the syslogs to see if we installed unknown firmware. However
with the omaha updater we don't have a way to stage the update but
not reboot, so we have to handle the device restarting out from under
us. While we did catch the ErrExitMissing, the connection could still
be in the process of shutting down, which could cause an error.

This refactors the code to also wait for us to receive a disconnection
event before continuing.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/950633
Original-Revision: 463d06f593c1b14b159668028d45e0f6f174fd95
GitOrigin-RevId: 8dc01c5313024c4fa019eb46c1fac00133dd4d72
Change-Id: I691fb7115761713d850d1b242344a929f171a2bd