[roll] Roll fuchsia [vim3][uart] create dfv2 variant of uart driver
Adds a dfv2 variant of the aml-uart driver and enables it on the vim3.

Tests: Created a copy of the dfv1 tests to work with the dfv2 driver.
Tested both v1 and v2 variants on a vim3 manually and confirmed working.

Original-Bug: b/301001701

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/948588
Original-Revision: e58376d05ae531c6226d4bb19f1e9473004dde18
GitOrigin-RevId: e07be8e6fa1ef904a28afd5a2bd2cdc8473feb61
Change-Id: I8b4096499e0c3cc35a54be1631a0fa45f35f25c4