[roll] Roll fuchsia [ffx][sdk] Don't resolve symlinks in sdk search
When run in google3 or otherwise through symlinks (such as in the bazel
sdk built in-tree), if we're searching for the SDK from the executable's
location, we should not resolve symlinks which means we can't use
`std::env::current_exe()` (which resolves them on linux but not other
platforms) or `Path::canonicalize()` (which always resolves symlinks).

In google3 this is especially problematic because the actual location of
the binary is in a CAS and not adjacent to the SDK at all, so it just
fails.

Instead, use argv[0] joined to the current working directory and search
from there.

Original-Fixed: b/312491133
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/950018
Original-Revision: 8b0498033261194094a625b9f43f7b182d4dfd76
GitOrigin-RevId: 6dde325a993f306836f7a08d175240bc1fa64c4b
Change-Id: I41b8e013cbf04c8e4820b0f5b2ab9ca3065e1897