[roll] Roll fuchsia [fxfs] Do not trim read sizes without supplying zeros.
Moves the size correction down after supplying zeroes for requested
regions outside the end of the file. The kernel shouldn't be asking for
those pages anyways, but there is a race with reads and truncation that
can result if this case isn't handled which will result in the pager not
fulfulling the entire request from the kernel and letting it hang.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/950643
Original-Revision: d76ec6a3351f9f69f588792b719f9a2850a9de98
GitOrigin-RevId: dcb5337a4b23a6bcdef7b4754c97eb5618b0fb75
Change-Id: I18da77a319178c904813c53d63e6aa84a8a8656d