[roll] Roll fuchsia [netcfg] Add stable name test for non-retry names
In netcfg interface name generation, if only a static name is specified
in the configuration naming policy, there is no capacity to re-try to
generate a new name to produce a different result.

Test that when two devices are added, and will attempt to use the same
name, that netcfg continues to work. For now, we can assert the name on
this interface, but in netcfg's upcoming change to remove temporary
names the test will update to ensure that the second interface either
causes an uninstallation of the first interface or denying the
installation of the second interface.

Linked to change made in fxr/947612

Original-Bug: 135087
Original-Bug: 56559
Original-Bug: b/309010918

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/948613
Original-Revision: 4c9aab77fd431c0d7cbe1813dbcd9ce963a16585
GitOrigin-RevId: 259607a394c6d71302233397e9c88117a9188957
Change-Id: I93f43aee40a059cf486d016bdf4e53bf0d26f893