[roll] Roll fuchsia [ffx][ssh] Adding sshconfig object API
This creates a struct that holds the sshconfig that is used
when building commandlines for ffx and subtools.

The struct is initialized from a sshconfig file which is compiled
into the library, but is usable directly by ssh using the `-F`
option. This has the advantage of being able to debug and test
configuration changes directly. The file can also be used by
non-ffx programs which are connecting to a fuchsia device via ssh.

The configuration can also be overridden programmically if the
need arises to change a default value.

The default configuration contained in this CL contains the
various configuration values that have been proven to be
an effective and efficient default configuration for connecting
to a device running Fuchsia.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/950031
Original-Revision: 8c907a7c5e323007da779ba24cb33ec56801c5d3
GitOrigin-RevId: 9f01f022080951655b58bfeb0a669628ac033967
Change-Id: I3f61250609a828ab561fcbaf5e5f1e709fb980d8