[roll] Roll fuchsia [assembly] Zedboot can be built when unused
We are moving in the direction of building the platform
once-per-architecture and then the product consuming ALL the platform
artifacts and building whichever product variant they wish. We need to
be able to build and export Zedboot to Bazel even when it is unused.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/951955
Original-Revision: 4d1cb77b634c55af763ee3066bd88a0a319523aa
GitOrigin-RevId: bf7fe0f4530c05c92af7351d867f9c306d552da7
Change-Id: I1e7162e68eaf00f69acb12c740f141c479283eab