[roll] Roll fuchsia [magma] Use structured config with magma conformance tests
The tests currently use command-line options to determine which GPU to
use in multi-GPU systems. Switch to using structured config to specify
that, as it will allow us to package the entire test as a subpackage
(possibly a prebuilt for OOT MSD development) and set the config from
the parent component.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/950735
Original-Revision: 7837a2781750f2b4d35cf46b0cd05c927555bee6
GitOrigin-RevId: cbeea1cb8ef05acaa829e64265300b1c4d927ae9
Change-Id: Ie5509c7f5f8f19a49897ea70528f3e386771223d