[roll] Roll fuchsia [testing][virtio_gpu] refactor and enable virtio gpu v2 unit tests
This change leaves a few TODOs inlined in order to keep the CL size
small. Namely, the test still needs to:
- Connect a fake sysmem using DFV2 testing best practices
- Define and connect a fake pci parent for the driver device
- Connect a fake gpu backend for the driver device

One departure from V1 of note is that the V2 implementation does not
support SetAndInitSysmemForTesting, nor creating a device with a
FakeGpuBackend, since the driver testing guidance encourages connecting
fakes through the test environment rather than through overwriting
internal properties of the driver.

Because of that, this change first
enables the tests now with real sysmem and a real (failing to connect
and erroring with ZX_PEER_CLOSED) bti and backend, with the intention to
fast follow up with fakes.

Original-Bug: 134883
Test: fx test virtio-gpu-test
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/952014
Original-Revision: 5272d035adb95f7321921dd4869c354e2179f582
GitOrigin-RevId: 580404ccb38df61ab7c8387f4ed11bb16d691e53
Change-Id: Ic40c8b4cda49349c6f6600932ac2d44f67992c52