[roll] Roll fuchsia [fidl][hdmi] Delete fuchsia.hardware.hdmi FIDL library.
fuchsia.hardware.hdmi/Hdmi protocol and other related types in the
same library are very specific to the amlogic-display HDMI video
output hardware components; it doesn't have a generic interface for
HDMI hardware configurations. Thus we decided in fxbug.dev/134691
that we'll delete this library.

Now that all the drivers have migrated out of this library, it's
safe to delete the library itself and all the helpers for types
in this library.

Test: no-op.
Original-Fixed: 134691
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/945638
Original-Revision: d2e53fab4026ce8e2d50fcc0ec268090707e6cb3
GitOrigin-RevId: a14097ddbd5b1a28f53e4e02018bdc314ff86037
Change-Id: Iad30547ab44a69c8527dffc24140b7db8a20e066