[roll] Roll fuchsia [amlogic-display] Merge the amlogic-hdmitx library into the amlogic-display driver.
The amlogic-hdmitx library is used for the soft transition from
the aml-hdmi driver to the amlogic-display driver, and it's
expected to have the amlogic-display driver as its only client.

Now that the transition is done, we can merge the HdmiTransmitter
class to the amlogic-display driver.

Test: amlogic-hdmitx-test moved to amlogic-display-tests
Test: no-op.
Original-Bug: 134691
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/945639
Original-Revision: a99fd079547d7ea4e79a18d44bd8be2d7ac56b38
GitOrigin-RevId: 899a5abd15ba4672f2c299241c5ec1cdf1224c97
Change-Id: I483a197976f395d758d94c416301f3768a4daa93